eilmer: Fix installation so all cfcfd3 dependency is gone.
With this commit, there should no longer be any dependency
on files from the cfcfd3 repository.

As part of the separation, we now build a specially named
lua executable 'dgd-lua' and install alongside the other
executables. We use this in place of the regular lua
executable so that we don't clash with any system-installed
lua. As a consequence, we now need to set two environment
variables explicitly: DGD_LUA_PATH and DGD_LUA_CPATH.

The species database is stored in the source area. Users
should also set the DGD_SRC variable to point to the top
of the dgd source area.

These new environment variables along with a typical install
proceudre are documented in the newly INSTALL file in the
eilmer area.