eilmer: reference solution is now dependent on time also
The reference solutions that may be provided for the
gas and solid domains (separately) are now evalulated
as a function of time as well as spatial location.
The function signature, provided by the user, is:

refFunc(t, x, y, z)

TODO: Update manufactured solutions example so that
it works with these changes to the post-processor.