kinetics: Added AlphaQssStep unit test.
The unit test checks that the order of accuracy for the AlphaQssStep
is in fact 2nd order. It does this by solving a chemical kinetics
problem over the same interval twice, with a different timestep on
each attempt. We can then check that the error drops as the 2nd power
of the timestep reduction.