gas: Changed GasModel constructor to be consistent with other Lua constructors.
While writing the documention for the gas package, the GasModel constructor
stuck out like a sore thumb (in that it wasn't consistent with everything elsewhere).

The old constructor took a single string (a filename) as argument:
   gm = GasModel:new('my-gas-file.lua')

The new constructor follows the pattern of always accepting a table.
Now the table contains one element, a single string:
   gm = GasModel:new{'my-gas-file.lua'}

Yes, this all just window dressing. But it avoids having to explain
the syntax differences to a new user in the *very* *first* example.