gas: Fixed molef2massf and conc2massf Lua functions to use new table arrangement.
I left myself a TODO last night to update the molef and conc tables
in the Lua wrapped functions to behave like the massf tables.
The idea being that values in those tables are set as "sp=val"
and missing species are set to 0.0. I've made that fix and removed
the TODO.