eilmer: Lua-wrapped FlowState will no longer accept invalid fields
I added this to help clumsy users. Chief amongst those clumsy users
is myself. I had erroneously set "u=someValue" expecting a very large
velocity in the x-direction. However, the code only accepts "velx" as
the field name for x-velocity. The new code will detect this problem
of invalid fields and report an error to the user.

This is all in the name of reducing the surprises for new users.