Eilmer: added infrastructure of postConvFluxAction bc.
We had orinally planned to have a (possible) series of
post convective flux actions applied as b.c. corrections
to the convective flux. This has not been needed until
now, so had remained as inactive (commented out) code.
Now, we have that need, so that code infrastructure has
been added.

A new InFlowBC_ConstFlux is added, but not yet tested.