eilmer: Make order of the cells reference list with the content of the files.
With a clearly defined order for the cells list in the unstructured block,
we should be able to rely upon it.  This might make the IO methods simpler
and it might make it simpler to implement the MappedCell boundary condition
such that it will be able to work for structured and unstructured grids alike,
since we'll be able to used a single index for cells in either type of grid.

It will certainly make the integration of unstructured grids into
the FlowSolution class a bit easier.  That's our first application.