eilmer: actually store references/pointers to the position vectors, as intended in SBlock::store_references_for_derivative_calc().
In the cloud of points for each vertex, I had been storing copies of the values
of the position vectors.  This worked accidentally until I moved the call to this
method to a point before all of the positions had correctly set values.
When dealing with the moving grid, we must be careful to set the individual
components of the Vector3 positions rather than replacing the whole object
(and this invalidating these stored pointers).