eilmer: don't let defaultThreadPool exceed nBlocks-1
I was still experiencing random segmentation faults when
using th parallel version of eilmer. It seems this was
only happening when the program had more CPUs to use than
blocks. I can only speculate about the cause for this.
A solution is to limit the defaultThreadPool to the smaller
of: maxCPUs-1 or nBlocks-1.

In other words, we don't need *more* worker threads available
than blocks to work on.