eilmer: Beginnings of a GPU chemistry module
This commit begins the work of porting Kyle Damm's
implementation of a chemistry integrator on GPGPU accelerators.

This commit introduces a dependency on the dlang opencl.d package.
There are several opencl bindings for D. The one chosen here
is a simple exposure of the C API by converting the headers
to D format. This package was obtained from:

  https://github.com/YusukeSuzuki/opencl.d