eilmer: Fixed moving grid bug in simcore.d, added shockfitting functionality to the moving grid implementation.
In the gas dynamic code for moving grid (simcored.d) the gtl was not being updated from 0 to 1 before performing the flow update, this update fixes the issue. The shockfitting implementation is also added to eilmerD. This update spans the four files listed. Currently shockfitting is known to be functional for ideal air simulations using one block with a moving west boundary.