eilmer: introduced some more explicit config flags for boundary conditions.
Various parts of the code, such as the reconstruction and flux calculations,
need to be aware of some of internal details of a boundary conditions.
These flags are included to make some of the testing easier than trying to
infer the nature of a boundary conditions from its lists of actions.

There are default values of these flags when a new boundary condition is
created in the Lua code.  These values may be over-ridden, as needed, in each
particular type of boundary condition.