eilmer: luaFlowSolution:get_cell_data now accepts 'fmt' option.
The 'fmt' field can be one of 'Plotting' or 'FlowState'. The 'Plotting'
option is the original table format prior to this change. If the
'FlowState' option is used, it in fact augments the plotting table
in place such that all valid fields for a FlowState table are in place.
The result of this can be passed to:

FlowState:fromTable()

Example code (fs:FlowState object, fsol:FlowSolution object):

   cell = fsol:get_cell_data{ib=0, i=345, fmt='FlowState'}
   fs:fromTable(cell)