eilmer: added 1st order one-sided reconstruction for shock fitting. Also added two step time integration compatibility for shock fitting. Finally, changed tcl shock fitting test Mach number reference erro allowance to 2%.
Shock fitting now has the ability to reconstruct the pressure, density and velocity values used to estimate the wave speed of the shock. It is also possible to run shock fitting simulations using predictor-corrector time stepping now. The shock fitting tcl test error allowance was changed from 1% to 2% for the Mach number variable as a result of 1st order reconstruction causing a failed test flag to be thrown.