eilmer: Re-connect time delay on starting chemical reactions.
The time sely for starting chemical reactions can be set from
an input script as:
  config.reaction_time_delay = 1.0e-5

Note I re-named reaction_time_start to reaction_time_delay to
be consistent with the naming of other parameters. The names
*_time_start and *_time_stop will be reserved for those parameters
that have both a window of time for their action (from "start" to "stop").