eilmer: Refactored shock-fitting code. Changed scaling factor to be applied to all vertex velocities along boundary. One sided reconstruction in shock-fitting code currently disabled.
Shock fitting code has been refactored. Parts of the code were rearranged and modified to adhere to DRY principle. Modifications to the reconstruction code appears to have caused a bug in the code whereby shock corrugation is noticed at the stagnation line. This push has the default interpolation setting for shock fitting set to 0th order whilst the bug is removed. The major algorithm change in this addition is the application of the scaling factor to all vertices on the inflow boundary, this smooths out the boundary by equally scaling shock-fitted and non-shock fitted vertices.