gas: Make species database more tolerant of missing viscosity or thermal conductivity data.
The species_data_converter program will now skip over missing items when trying to
convert an eilmer3 data file to eilmerD format.

The prep_gas program will now give a warning when it cannot find required data in
the database for a particular species. In that case, the value in defaults.lua will
be substituted.