kinetics: Set default 'max_attempts' to 4.
Making this default value 4 brings it in line with what
we had in eilmer3. I'm not sure why I set it to 3 when porting
over the new code. In any case, using 4 max_attempts seems
to give added robustness for hydrogen combustion cases.
The extra attempts at getting the chemistry correct are needed
right at the stagnation point of the Lehr combustion cases
during the initial transients. This occurs right as the shock
is first reflecting from the body. It makes sense that several
attempts are needed to get the chemistry to act stably: there is
a lot of flow property change happening in a short period time
as that shock reflects.