Eilmer: high-order reconstruction for convective terms working for USG.
We needed to handle the degenerage case where there was only a single line
of cells and these did not align with the interface normal.  This resulted in
linearly dependent rows, in that one was just a factor multiplied by the other.

We also needed to apply clipping of the reconstructed values.
No real surprise there.

Anyway, that's the good news.  The bad news is that it is dreadfully slow
to compile and very slow to run, especially with multiple cpus.