eilmer: Turned reconstruction back on in shockfitting. Added in an optional         1st order reconstruction which can be turned on in the source file.
After reviewing the code in light of the stagnation line corrugation
I am of the current opinion that there is no bug. Instead the
algorithm is sensitive to both inflow boundary definition and
CFL number. Both should be carefully chosen to ensure 1. the
initial grid somewhat conforms to the expected shock shape, and
2. the CFL number is small enough considering the final grid.