kinetics: Brought chemistry update method more in line with what we had in eilmer3.
I had had difficulty getting the Lehr validation case to work as well as it
had in eilmer3. There were two key pieces missing from the chemistry update
algorithm:

1. A check on mass fraction error.
We do two checks at then end of an ODE step. A check that the step passed
based on the ODE method itself, and check based on physics. That check based
on physics is a mass balance check. I had neglected to add this check back in.
Along with that, we do a scaling of mass fractions at the end of each global
chemistry step.

2. Timestep reduction on failure.
I had gone with Mott's suggestion of a reduction in timestep by a factor of two
or three when the steps failed. This is not aggressive enough for our cases and
leaves us spinning our wheels while we do multiple attempts and only modestly
reduce the timestep on each failed attempt. I've gone back to what was tried and
tested in eilmer3 over many years: a reduction by a factor of 10 for a failed chemistry
step. I suspect Mott's experience that led to a suggestion of two or three was NOT
in high-speed compressible flows. I suspect the shocks that rattle around our flow
domain warrant a drastic drop in timestep when things go awry.