eilmer: Added user convenience function: makeFillConditionFn
This function can be called by the user to set up a function that
can be supplied as a fillCondition for a Block object. The function
accepts a FlowSolution object and returns a function. That function
can be supplied as a fillCondition object. As en example:

fsol = FlowSolution:new{jobName="myJob", dir="../coarse",
                        tindx="last", nBlocks=4}
initial = makeFillConditionFn(fsol)