eilmer: Implemented new point cloud stencil for unstructred grid         spatial derivative calculations.
A degradation to sub-1st order convergence was documented for
multi-block N-S unstructured MMS simulatuions. It was theorized that
this drop in order was a result of non-symmetric boundary clouds
which resulted in boundary interfaces having two different spatial
derivative estimates dependant on which block was active.

This issue was resolved by implementing a symmetric cloud stencil which
returns the same spatial derivative estimate at boundary interfaces
independant of which block is active. The fix has provided 2nd order
convergence even for multi-blocked N-S simualtions.