eilmer: Rework the Muses-C capsule in X2 facility simulation.
We've been using a plot from this very old simulation as our advertising image
for a lot of years.  It really needed refreshing, however, the simulation was
done so long ago (in the 1990s), the input file was for the original C-coded
interpreter scriptit.  The original input file was recovered from a backup CD
written in 1998.  It is the file cap.sit.