Eilmer: reduced repeated code by using some mixin statements.
The code for setting of fields in the GlobalConfig class from values in the
Lua table was very repetitive and annoying to update.  These commit cuts the
number of code lines by half and makes it easier to scan the items.

Of course, the conservation law of evil causes the remaining lines to be much
longer.  I'm prepared to tolerate those long lines because the full collection
is so much easier to scroll through (on a sufficiently wide editing window).

There should be no change to function with this commit.