Eilmer: fixed attempt to write all digits for floating-point numbers.
The problem with the previous commit was that the JSON reader would refuse to
extract double (floating-point) values if the string representation that was
read did not have a decimal point.  The %.18g would write integral values
without that required decimal point, so we've moved to using %.18e as our
floating-point number format to write all digits of the fractional part.