eilmer: Adds weighting to least-squares viscous gradient code.
The Hakkinen case was throwing bad cell errors when using the
least-squares gradient estimation at the interfacial positions
for the structured grid solver.

A quick literature review found that the norm is to use weighted
least squares for the viscous gradient estimation where the weighting
is based on the inverse distance.

This update adds weightings directly to the 2D least-squares
gradient calculation. It is noted that currently this implementation
is working for the structured solver (the Hakkinen case now evaluates
the correct solution using least-squares at faces)
and the unstructured solver (with no ghost cell data in the
viscous clouds). Currently a bug exists for the use of ghost cell
data in the viscous clouds.

ToDo: separate weightings from flowgradients methods to reduce
      computational redundancies (i.e. only evaluate weightings
      once at simulation start and/or when the grid is moved)