eilmer: Completed least-squares implementation.
This revision builds on the previous weighted
least-squares update r546. A number of major & minor changes
have been performed to accomodate the new implementation.

1. Structured grid viscous point clouds for the "faces" spatial
location now store the interface in the first cloud position.
(Note currently this modification breaks the cht-MMS test case,
the reference norms will need updating).


2. The calculation of the least-squares viscous gradients about mid-points
has been removed for the "faces" spatial location. It was found that for
the sparse clouds required at the boundaries for the unstructured grid
to remain 2nd order for multi-block simulations, the mid-point value was landing
too far from the interface point which resulted in bad cells appearing.
It was decided that it is more direct to just calculate the gradients at the point of interest,
i.e. the interface position. For the "vertices" spatial location this is not possibe since
we do not store flowstate data at vertices, thus the mid-point method is still retained
for this case. (Note this modification breaks the NS-MMS-lsq test case, the norms will
need updating).

3. Least-squares weights are no longer calculated in the gradient method. There is
a method housed in flowgradients.d that now calculates these. We need only calculate
the weights at the beginning of a simulation and if the grid is moved.