eilmer: Update to sblock viscous cloud storage.
The latest weighted least-squares update (r551) broke the
divergence theorem calculation. We had to modify the structured
viscous cloud stencil to accommodate the weighted least squares
method. However, the divergence theorem calculation expects
the coulds to be stored in a specific order. This update allows
the code to do either depdending which spatial derivative calculator
is chosen.

Also added in this update is the modified reference values for the
structured NS-MMS test case. All tests should pass now (and do so on
my workstation machine).