Eilmer: refactor least-squares reconstruction to reuse normal-matrix inverse.
This refactor, to retain the normal-matrix inverse and the other intermediate
data within the FVInterface objects, allows the simulation to run close to the
structured-grid speed (29 seconds vs 25 seconds cor cone20-xxxx).

A first cut was made by Kyle to show the speed improvement for the
unstructured-grid calculation.  This commit goes a bit further to reduce
code repetition and to make it ready to optionally have the workspaces
stored in the FVInterface objects, as they are for this commit, or to conserve
some memory and use a common pair of workspaces for all interfaces in a block.