Eilmer: add config option retain_least_squares_work_data.
Default is true.  The lsqinterp workspaces will be retained within each
FVInterface object and the inverted normal matrix reused in resonstructions.

If you happen to run out of RAM for a particularly large simulation, try
setting the option to false to not allocate all of the local workspaces.

So far, it is only relevant for unstructured-grid convective-flow reconstruction.
We should do something similar for the flow gradient calculations for the
viscous terms.