eilmer: added 2 new unstructured gradient limiters.
The min-mod and van albada limiters appear to work fine for
low Mach no. flows on unstructured grids, but fail to solve
simulations for high Mach no. flows. These limiters generate
bad cells for Mach 14 flow 20 degree cone.

This commit contains two new limiters developed especially for
unstructured grids. The first (Barth Limiter) is considered the
first limiter for unstructured grids, it exhibits monotonicity but
loses 2nd order convergence in the presence of discontinuities in the
solution domain. The second (Venkatakrishan Limiter) is a modified
Barh Limiter, it is a smooth limiter which *apparently* holds second
order convergence but at the cost of not exhibiting monotonicity in the
presence of large gradients (shocks). This can cause small oscillations
in the solution domain near strong disconintuities.

Both the new limiters are able to converge on the solution for the test
simulation of Mach 14 flow over a 0 degree cone.

NB. This push temporarily breaks the 3D sod-usg test case.