Eilmer: have a simple version of second-order reconstruction back together.
The issue with the previous commit was that I forgot to avoid trying to using
the gradient data for the ghost cells.

This particular commit goes back to the simpler (but not so great) component
by component limiting across the interface to which we are constructing flow
data estimates.  Kyle will replace this once again with his better limiters.