Eilmer: clean up arguments for constructing vertices and faces.
We messed up on passing through the flags for allocating the least-squares
workspaces.  For structured grids, the computed boolean flag was being passed
in as the integer that initialized the id.  Presumably the id was initialised
to either 1 or 0, depending on the boolean value, however, this wasn't noticed
because the id value was almost immediately overwritten with the correct value.

How did we get into this mess?  Well, (1) I changed the arguments to the
constructors and didn't change all calling locations consistently.
(2) Had one mostly-unused argument with a default value, and ended up
with the misaligned boolean being silently cast to an integer by
the function call (I think). (3) Had a trailing assignment assign that integer
id value after construction and effectively cover up that part of the problem.

What a sad tale of poor programming practice.