eilmer: Added in message for lsq gradient workspace.
The least-squares gradient workspace is initialised in the fvcell.d
script however the workspaces are instantiated in ublock.d this is
because we don't use the workspaces in the structured solver. I added
in a message describing this for anyone who might be puzzled as to why
we don't instantiate the workspaces where we initialise them.