eilmer: Updated K variable and fixed a bug in Venkat limiter.
I had accidentally pushed the copy of the Venkat limiter which had
the tunable parameter K set to 100 instead of the commonly used value
of 0.3 cited by Venkatakrishan.

I also amended a small error where in some instances the phi limiting
factor was not being set correctly in areas with small gradients.