Eilmer: have enough plumbing in place to import a 2D grid in su2 format.
We can pick up the su2 grid, prepare an initial solution and postprocess that
initial solution to produce a VTK plot file.  The example cone20 has be prepared
by Kyle using Pointwise.  It uses only triangular cells.

Haven't yet tried to run the simulation because it's lunch time and I want to
enjoy the success of getting this far for an hour or two.