Eilmer: make flow gradients available in ghost cells for usg-reconstruction.
We would like to the best gradient estimates in the ghost cells, prior to
the high-order reconstruction of the flow field on unstructured grids.
For ghost cells along boundaries that are mapped to other cells,
the gradient data is available and should be copied.
For ghost cells that don't have a mapped cell from which to pull the gradient
data, fall back to just copying the data from the cell on the other side of
the attached face.

To be sure that we copy the most recent gradient data, we have split the
convective_flux method into two phases.
The need to dipping into data that had been hidden in the boundary
condition classes required adding some extra support methods to return
references to the mapped cells.

Finally, we changed the offset for id values of ghost cells to one million.
After Sangdi's set up of a 3D wedge simulation with a million cells, the old
offset value of 100,000 seemed no longer large enough.

This commit finishes the job started in rev. 656.