Eilmer: implemented the viscous_delay feature.
We may choose to start a simulation with the viscous terms suppressed by setting
config.viscous_delay to something larger than 0.0.
If we do this then the simulation starts with viscous_factor=0.0 and will start
incrementing viscous_factor once sim_time exceds viscous_delay.
The viscous_factor will be incremented up to 1.0 in steps of
config.viscous_factor_increment, which is set to 0.01, presently.