eilmer/examples/sawada: Made testing friendlier for those without dgd in $HOME
This commit introduces the environment variable DGD_REPO. This environment
variable can be used to set the location of the dgd repository.

For sphere-sawada case, this allows the test to find the look-up table
if the user has not installed the repository directly under $HOME.
I separate the prep and run stages into two steps so that I could
do the detection of the DGD_REPO variable in the prep stage.