Eilmer: fix bug in code for specifying subgrids in SBlockArray.
If we received a grid with a number of cells that could not be divided into
equal chunks, we would subdivide correctly only on the first pass, changing
the increment of cells on the last block in a row or column.  We then neglected
to change this increment back to the nominal value when starting the next row
or column of subblocks.  Fixed by using local names and not changing the
nominal increments.

Tested only on Kyle's cylinder-flare case.