Eilmer: moved stringent_cfl and viscous_signal_factor to .control file.
Aligned .control parameters as listed in prep.lua and the reading function in D.
Also update the relevant parameters after rereading the .control file at each
time step.

The manufactured solution tests now fail the check on the number of steps taken.
Need to check why, but I suspect that stringent_cfl was not being correctly read
in recent times.