eilmer: Added a pre-iteration stage to the steady-state solver.
When we start with constant initial conditions everywhere
(which is a common choice), this can be a poor initial residual
estimate for the steady state solver. In this case, the initial
residual has zero contribution from the interior and only a
contribution from boundary conditions. We use the drop in relative
residual to control the switching between pseudo-transient continuation
and the inexact Newton phase. If we have a poor guess for initial
residual, then that has influence on our relative residual.

In this commit, I've added a pre-iteration phase. It takes a small
number of steps (controlled by nPreIterations) in which to establish
a maximum residual value. This allows the boundary conditions time
to exert some influence on the interior. During this phase, the
timestep is kept constant and only low order reconstruction is used.
The maximum residual is then used to compute the drop in relative
residual values as the iterations are progressed.