eilmer: wrap the idealgasflow module for use in simulation preparation.
Just enough has been wrapped (beta_obl) so that the Billig correlation functions
could be used in the building of the geometry for Troy's 3D cylinder simulation.
There, we want to build a curve approximating a shock over the cylinder so that
the resulting grid is somewhat aligned with the actual chock.

The functions have been put into a table called idealgasflow so that the specific
function may be called as:
local beta = idealgasflow.beta_obl(M, theta)

We expect to find more uses for these functions so they are available in the
preparation and custom-postprocessing modes of e4shared.