Eilmer: for 2D structured grid, make consistent order of vertex list for interfaces.
When following a 2D interface (line) from first vertex to last, the unit normal
for the face is to the right, pointing into the right-hand cell.  Since the
face geometry has always been constructed via the global indices, there is no
change to face normals and right-hand cells; just to the order of vertices
reference by the list.  We may come to use this list in future and depend on
the order being correct, now that the code is mostly an unstructured-grid code.
At least this fix should eliminate one source of confusion when debugging.

Also, set lengths of dynamic arrays to zero before appending to the lists.
Although the lengths should have been zero already (from initialization),
this makes it explicit.