Eilmer: do the detection and recovery of bad cell data a bit more carefully.
Previously, the decode_conserved function was throwing and exception too soon
and not giving the count_invalid_cells function a chance to patch things, if
appropriate.  Now, a flag in the cell is set if the thermo data has gone bad.
Also, count_invalid_cells gets a chance to check this flag after every stage
of the gasdynamic updates.