bc: ghost cells now make use of the mapped cell position data.
Ghost cells that are mapped to finite-volume cells along a block
connection interface now use the mapped cells position.

This is important for the least-squares inviscid flux reconstruction,
where the gradients from the mapped cells are stored at ghost cells to
ensure consistent fluxes at shared interfaces. However, complete consistency is
only possible if the gradients of the ghost cells are multiplied by the same
distance from the mapped cells center to the interface mid-point. By making
use of the mapped cell centre position, we achieve this.