eilmer: Upgraded iterative solver to a restarted flexible GMRES method with preconditioning.
The Newton-Krylov update now handles 2D laminar flows on a single block.
Preliminary tests with the basic GMRES algorithm showed rather slow
convergence. A simple unpreconditioned GMRES method worked rather well
for 2D invisicd flows, however, that didn't carry over well to a
laminar flate plate test case at Mach 4.

I have added several upgrades to the algorithm in order to help:

1. Option to use a preconditioner.
I have chosen to use the GMRES iterative method itself as preconditioner.
This leads to what some have called an inner-outer method. When using
a varying preconditioner (such as an iterative method), the regular
GMRES method is no longer applicable. As such a right-preconditioned
flexible GMRES method is used in its place. The 'flexible' term refers
to the fact that there is flexibility in the choice of preconditioner
between iterations of the main GMRES method.

2. The "restarted" variant of fGMRES is implemented.
In some cases, the GMRES method requires a very large number
of iterations (and a correspondingly larger number of Krylov vectors)
in order to achieve a satisfactory level of convergence.
Unfortunately, a large number of Krylov vectors leads to large
memory requirements. The other drawback is that the orthogonolisation
process must orthogonalise against all previous Krylov vectors during
iteration. This leads to a large expense in arithmetic as the number
of Krylov vectors grow.

A remedy to this is to cap the number of Krylov vectors at some
modest amount (say 30-40). If the iterations have not converged
at the end of using all those vectors, then take the best estimate
of the solution and start again with that as the initial guess.
This method should still converge monotonically, but some
convergence optimality is sacrificed in the restart process.
However, it does alleviate the memory and computational
expense of a large Krylov dimension. The number of restarts is
capped with the parameter 'maxRestarts'.