Eilmer: now that we can collect cells along a line, add the shock angle tests
...for the sharp-cone, unstructured-grid cases.

Unfortunately, the fast version of the find_enclosing_cell method is not working
reliably for picking up the points along a line.  We see info messages saying
that it cannot find cells in places where we are pretty sure that cells should
be.  For the moment, the slow version of the method is used.  Look for the FIXME
hack on lines 271 and 271 in src/geom/grid.d.