grid_utils: Fixed bug in local node ids for ugrid_partition.d
The initial (naive) approach to transforming global node ids to
local node ids for the partitioned su2 files was a failed effort in
keeping things simple. In this update each block has a transform
dictionary that stores the local node ids which can be retrieved
by using the global node id as a reference, i.e. transform_list[global_id].