grid_utils: ugrid partitioner now writes out mapped cells file.
This is the first update for allowing mapped cells information to be
passed between the ugrid partitioner and eilmer4. This commit updates the
ugrid_partitioner program to writeout a mapped_cells file which is in the
format:

block_id local_cell_id neighbour_block_id local_neighbour_cell_id

NB. The file explicitly lists all combinations of mapped cells i.e.
each shared face is represented twice, for example,
0 10 1 23
and
1 23 0 10
are both contained in the file.