Eilmer: Allow separate settings for perturbation parameter in different phases.
sigma0 : perturbation parameter during start-up phase
simga1 : perturbation parameter during inexact Newton phase

Since we typically use a low-order Jacobian early on, an appropriate
value for sigma in that phase can be quite different compared to one we'd
like to use later on. Sigma is the parameter by which we perturb the
solution in order to form the finite-difference approximation to
the Jacobian. In choosing sigma there is obviously the competing effects
of choosing a value that's small enough to reduce truncation error,
but large enough to avoid significant round-off error.