Gas models: started the process of cleaning up notation
In preparation for the nonequilibrium thermochemistry.
The main change here is to rename some of the GasState members.
old   new        comment
-----------------------------------------------------------------------
T[0]  Ttr        scalar value for trans-rotational temperature of the gas
e[0]  u          scalar value for thermal energy of the gas
k[0]  kth        scalar value for the thermal conductivity of the gas
T[]   T_modes[]  array of temperatures for internal noneq energy modes
e[]   e_modes[]  array of internal energies for thermo nonequilibrium
k[]   k_modes[]  array of thermal conductivities for the internal modes

This commit is a part way point for me (PJ).  It is not ready to be used.
The thermal conductivity tests are broken and the gas-dynamic code is yet
to be adjusted to this new notation.