Eilmer: have now changed to the new thermo notation introduces in rev 825.
At this commit, we have a clean build of the codes and the cone20.test case in
examples/eilmer/2D/sharp-cone-20-degrees/sg/ runs without complaint.

I'm going to declare this an adequate victory for the week and retire for the
night.  It is known that a few scripts need to be updated because we now need
to refer to the temperature as a scalar in the Lua scripts.