Gas models and Eilmer: eliminated the moderately ugly kth in favour of k.
We temporarily used kth as the name of thermal conductivity for the gas while
we made the changes to the internal energy and modal temperature arrays.
The old array of thermal conductivities was called k and we wanted to be sure
that we cleaned out all occurrences of that variable before reusing the name k.

Also, in this commit, we have cleaned out a few more occurrences of T[0] in the
example scripts.