eilmer: Removed some redundant code from lsqinterp.d which causes inconsistent fluxes
In the least-squares convective flux update stage, we were checking whether
a cell had a work-space allocated or not. If a work-space wasn't allocated,
then the face-neighbour cell's limiter value was used during the reconstruction.
This was causing inconsistent fluxes at shared interfaces along block connections.
Since the limiter values are stored at cell centers, and we communicate this
information in the ghost cell data, we are always able to use the correct limiting
value (except when at the edge of the domain, however this is handled in ublock.d).

This upate removes the redundant work-space check when applying limiting in the
reconstruction stage.