examples: Tweaked Kyle's metis test script to first partition the mesh
and added the test to the top-level script.
It takes only a couple of minutes on my little HP laptop.
I assume that anyone interested in running the automated tests is also keen
enough to install metis on their system, or edit the top-level script to
remove this particular test.