Gas: added the Powers-Aslam gas model and fixed makefile
...to cope with the tighter coupling of the gas models to Rowan's kinetics files.
Now that there is a function to make a ThermochemicalReactor object in the
base gas_model.d file, we need to include the kinetics files in the compilation
process of the gas models files.

The Powers-Aslam gas model is partly a demonstration of how to add a very
specific gas model that has a tightly-coupled reaction scheme that is coded
directly into the gas model file.  It also makes a good validation example
for the flow solver.