gas: Make ceaSavedData available as a table entry in the Lua-wrapped GasState.
One could get a list of the mass fractions in a CEA gas from the Lua
interface using the following code snippet:

for sp, mf in pairs(Q.ceaSavedData.massf) do
   print(string.format("massf[%s]= %.4e", sp, mf))
end

Also, I changed all of the internal initialisations of the GasState to use
the constructor that takes a single GasModel object. This constructor takes
care of the decision to test for the CEA gas and initialise the CEASaveData
struct if required.