gas: Add the Wilke mixing rule for viscosity and thermal conductivity to gas tests.
This completed the band-aid job I did in rev 915. At that stage, I just commented
out the test completely so that we could get on with some other work.

The tests for the Wilke mixing rule are now active again and have been correctly
plumbed into the automated test script.