gas: Lua interface now returns molMasses table with species names as keys.
While I was fixing up the indexing inconsistency in the Lua-wrapped speciesName()
method, I also notes that the molMasses() method return an array numbered
from 1 with the molecular masses. It seemed the cleanest solution was to return
a table with species names as keys rather than an index.

In general, the user should interact by using species names in the Lua interface.
This will be the default interaction with tables of mass fractions, mole fractions,
molecular masses, etc. Only when there is compelling need will we interact with
species based on indices. When that's the case, we'll use a 0-offset even in Lua.