nm: Added a wrapper for the bare-bones linear algebra model.
I have exposed the basic functionality of the linear algebra module.
This is not a complete wrapping of the module but it should
expose enough functionality so that Shaowen Bai can make
use of the matrix solver in his work on moving grids.
Shaowen is using the Lua interface to do his moving grid work and
requires a matrix solver to compute the motion of the grid vertices.