eilmer: Added config flag 'coupling_with_solid_domains'.
Now there is the option as to how to handle the coupling
between solid and fluid domains. Everything in eilmer4 had
been tight coupling up until this point.

In this commit, only the tight coupling is hooked up.

In the next commit, we'll make a loose coupling option
available so that Nigel Chong can hook in his
bakward Euler update.