Eilmer: Added universe_blk_id to LocalConfig.
We wanted a way to get the identity of a block close
to a cell. A cell already has access to a LocalConfig
object so we can poke the universe_blk_id into that object.

The immediate motivation was so that we had more context
for error messages when things go bad at the cell level.
This was a suggestion of Ingo Jahn and Miguel Pengel.