Eilmer: User-defined grid motion starts with all vtxVel = 0.0.
Ingo Jahn and Miguel Pengel noted that non-moving blocks
required an explicit action of setting all vertex velocities
to 0.0. This was tedious and non-intuitive.

The fix is that now *all* vertices are set to 0.0 velocity
at the start of each step, behind the scenes. The user's
grid motion function then only needs to concentrate
on setting those vertice with non-zero values.