Eilmer: Added Lua helper function sampleFace().
It is something like sampleFlow (aka sampleCell), but with the signature:
sampleFace(which_face, blkId, i, j, k)
where the arguments are:
  which_face is a string "i", "j", "k", or "u"
  blkId is the index to the block
  i, j, k are indices of the face in a structured grid context.
The returned table contains contains a number of names entries,
specifying the face geometry and the flow state.

See functions pushFaceToTable() in lua_helper.d and pushFlowStateToTable()
in luaflowstate.d for details of what the entries are.

Beware, there is almost no argument checking at the moment.
Handle with care; there are likely to sharp edges.