Eilmer: moved the initialization of the Lua interpreter for a boundary-condition
into a function within the BoundayCondition class.
This seemed to be the clearer thing to do since the Lua interpreter is owned by
the BoundaryCondition object, not the individual Effect objects.

Individual Effect objects may want initialize the interpreter but only the first
will actually take effect.  This means that you should put all of your user-
defined effect functions into the one Lua script.

As well as shifting the init function, we have added some extra boundary
condition attributes to the global namespace of the interpreter.
These are label, group, type.

Also, added label to the infoFlowBlock helper function for Ingo.