Eilmer: Use find_enclosing_cell in makeFillFunction
There is a convenience function in prep.lua that can
create a 'fill' function based on an old solution.
That convenience function had used calls to
'find_nearest_cell_centre' to do its work. However,
I found that this would give a ladder effect on
grids with high aspect ratio (like those used for
supersonic/hypersonic boundary layers). Instead,
I changed to using calls to 'find_enclosing_cell'
and this seems to do a much nicer job when transferring
a solution from a coarse grid to a finer grid on
grids with high-aspect-ratio cells.