Eilmer: add a couple of config items for Kyle's loads output.
config.boundary_group_for_loads = "loads" -- We may change the group name.
config.compute_loads = true -- We may turn off the feature

With these default values in place, the feature behaves as is expected in old
examples but we now have a bit better control of when loads are computed.