Eilmer: Load the 'lua_helper' module behind the scenes for user-defined scripts.
This makes the 'lua_helper' module (which is pure Lua) available to
some of the user-defined hooks. Since it's a Lua module, we load
it using the Lua "require" function. There's no direct C API access
to that, so we dip into the Lua interpreter to set that up from D-land:

   luaL_dostring("require 'lua_helper'")