Eilmer: Fixed ausm_plus_up for k-omega work; put in guards against incompatible flux calculators
Presently there is a check in place for selecting flux calculators that
are incompatible with the k-omega model. The list of incompatible flux
calculators is:
  + efm
  + adaptive
  + hlle

If EFM can be fixed up to work with the k-omega model, then the
above list will reduce to one item: hlle.

Also, changed the selection of flux calculator in the Coles case to
ausmdv so that the example was not broken by this change.