eilmer: modified calculation of the h parameter in Venkat limiter.
Initially I was computing the volume (3D), or area (2D) using the average
cell length, for use in calculating the h parameter in the Venkat limiter.
I have changed this to use the cells actual volume, or area.

I also cleaned up the Venkat limiter code to be more aesthetically pleasing.