Eilmer/geom: Added Tecplot binary output to --post.
Thanks to Pierpaolo Toniato for working out the
details of using the Tecplot library for writing
the binary szplt files, and for writing a D wrapper
for the functions we need.

In this commit, I've incorporated Pierpaolo's code
so that the Tecplot binary output is available as
an option. It is the default when using "--tecplot".
To get the ASCII, use "--tecplot-ascii".

This new feature only works if the user has an installation
of Tecplot on their machine since it needs access to some
Tecplot-supplied shared libraries. There are some
new notes on setting this up in the INSTALL file.