Fixed deprecation warnings for DMD 2.075.
These were mainly using std.regex.replaceAll instead of std.string.removechars
and using -g instead of -gc when invoking the compiler.

There were another couple of small changes over in kinetics to get the tests
to build.  A couple of these fail with the test code seeming to be a bit stale.