eilmer: Beginnings of the CUDA GPU chemistry module
This commit begins the work of porting the GPU chemistry module over to CUDA,
from OpenCL.

This commit introduces a dependency on the dlang cuda.d package.
There are several cuda bindings for D. The one chosen here
is a simple exposure of the C API by converting the headers
to D format. This package was obtained from:

  https://github.com/prasunanand/cuda_d

Currently this dependency hasn't been linked into eilmer4, this revision simply
adds the cuda.d package to the repository, along with a small Hello World program,
for testing the package.