Eilmer: Added a wall_temperature_on_init parameter.
When using the special initialisation condition to diffuse
boundary values into the domain, the temperature at the wall
was not effective for adiabatic boundary conditions. This
is because the temperature is not directly computed in that
boundary condition, nor do we know what the temperature will
be at the wall until the flow is established.

This new parameter allows the user to give a reasonable guess
for a wall temperature to diffuse. A useful starting guess
might be to give the total temperature of the free stream flow.
In any case, direct control is given to the user when a finite
positive value is given to this parameter:

config.wall_temperature_on_init = 275.0 -- K