Gas: Added a two-temperature nitrogen model, with vib relaxation.
A two-temperature, single-species nitrogen (N2) model is added.
Accompanying it, is a thermochemical reactor subclass that
handles a Landau-Teller type vibrational relaxation. Presently,
the relaxation time based on Blackman's data is hard-coded.
This model allows us to do the 2-T version of the Giordano
test case.

TODO: Allow the relaxation time to be configured from a file.