gas: Modified the two_temperature_reacting_argon file to have chemistry and two temperature model. Model implementation to be checked by Rory Kelly.
Still requires a proper model for varying Cp and Cv, and also requires the
chemistry step to be implemented properly. At the moment it attempts to work out
the new values of mass fractions using a single step. Will investigate explicit
and implicit ways of doing this step so that the code does not crash.

Added in stefan boltzmann constant and avogadro's number to the
physical_constants.d