Eilmer: Allow fillCondition to directly accept a FlowSolution object.
I've pushed the call to makeFillConditionFn to behind-the-scenes.
This more closely follows what we accepted as a fill_condition
in Eilmer3.

There are still cases when the user might want to something more
exotic than simply supply an old flow solution. For example, he/she
might want to add extra species to the fill condition. This can
still be accommodated by constructing a bespoke fill condition function.