Eilmer: postpone flagging of boundary cells and faces until first BC application.
This means that we first attempt to flag the boundary cells and faces once
the actual cells and faces are constructed within the FlowBlock.
I had given Rowan poor advice on Thursday last when I suggested putting this
flagging action into the constructor for the boundary effects for Wilson's
Wall Function boundary conditions.  Sorry about that.