Eilmer: revisit profiling
Get rid of -vgc from the compiler flags; it's too noisy.

Introduce Vector3.clear() for setting zero components in the Vector3 objects.
This was part of a bit of fiddling around to try to optimize the method
clear_fluxes_of_conserved_quantities().  Athought there was no speed increase,
the code was a little tidier and, presumably, the optimizer needed to do less
work.

In simcore, init_simulation(), suspend the garbage collector.  This might have
some benefit for the unstructured-grid simulations, but I have yet to measure.