Eilmer: have once again allowed GC to be active during initialization.
With large calculations, it is a bad idea to disable the garbage collector for
the whole initialization because the memory requirement grows very large.
For Jen's nozzle example, a requirement of 6 to 7 GB increased to about 27 GB.
We need to consider finer-grain control of the GC activity, possibly suspending
it only in the functions that read and write individual blocks.