Eilmer: Use Hessenberg matrix in solid domain GMRES.
In this commit, we added the use of the rotated Hessenberg
matrix to the GMRES algorithm. This allows a "cheap" estimate
of the residual at each iteration of a GMRES step. If the
residual is below a desired tolerance, then the steps terminate
early and the solution vector is computed.

This differs from the standard GMRES algorithm that will just
step for a pre-determined number of iterations.