Eilmer/examples: Made test suite tolerant of missing gpmetis and clean-up earlier partitioning.
The cone20-usg-metis.test would fail if the user did not have gpmetis
available, or if paritioned blocks were left behind from a previous run
of the tests. Now, the test is skipped if there is no gpmetis on the
system. Also, the test will attempt to clean-up old partitioned blocks
at start-up.