Eilmer: to be compatible with 32-bit systems, we should use size_t consistently.
There were a couple of locations where ulong variables were declared and then
used to index arrays.  That is OK on a 64-bit system but is not reliable on a
32-bit system where the addresses are shorter.  It seems that recent versions
of DMD are checking more carefully and it has been a while since we have
compiled on a 32-bit machine, hence the need to change this fairly old code.