Eilmer: Added a limit on tke production related to local thermodynamic energy in cell.
There are two new configuration options that allow one to impose
a limit on the rate of production on turbulent kinetic energy.
That limit can be related to a local change in temperature,
in that, we restrict the change to be no more than a certain
amount of degrees Kelvin (default is 5.0). These new options
are set in an input file as:

config.limit_tke_production = true
config.tke_production_limit_in_kelvins = 5.0