gas: Change starting guess for iterative pT evaluation.
Sometimes we rely on an iterative approach to updating
the thermo state when given pT. In that iterative approach,
we make calls to the rhou function. We had a hard-coded
starting guess for that function that was not very robust
at high temperatures. In this new version, the pT function
will take values of rho and u inside the GasState class
if they are available. If not available, it will fall back
to those hard-coded initial guesses.