Eilmer/util: Added deepclone of tables and apply to bcList when constructing blocks.
At Fluid/SolidBlock construction phase, we need to take a copy of the
bcList table that user provides. This is because the user might declare
just one table and re-use that many times. If we then change
the bcList associated with just one block, it will affect all other
blocks that share that same table. This could (and did) happen
when changing boundary conditions during a block connection phase.

By taking a copy, we avoid the problem of shared references to
one table.