gas: Played with iteration settings again for pT call with LUT
In a few commits back, I tried using a more sophisticated
selection for the initial condition. This only helped robustness
in some situations.

In this commit, I've gone back to the hard-coded initial guess
but I've increased the number of iterations allowed. Tim Cullen
made this suggestion. I've increased from 20 to 30 and this seems
to do the trick.