Geom: Gather orphan faces when importing SU2 grid.
These faces are collected and put into an extra broundary set so that
a suitable boundary condition may be applied and the simulation proceed.
The ExchangeBC_MappedCell may be a suitable BC to use to connect cells
that aren't quite matched at the orphan faces but are probably intended to
be matched.

Also, had to finally implement the inside_xxx functions for pyramid-
and wedge-cells.