Eilmer: fix the west-boundary flavour of the subsonic inflow, from stagnation BC.
The essential change is to make the increment in pressure sensitive to
the direction of mass-flow through the boundary.  Also, if the mass-flow is out,
set the speed in the ghost cell to zero and apply full stagnation conditions.

These changes are motivated by the observation that sufficiently strong
subsonic inflows have behaved well but the very slow flows driven by small
pressures (as seen in the microcombustor studies by Xin and Britt) have often
misbehaved.  Once the flow had been nudged the wrong way (i.e. outflow),
the boundary condition made things worse.