Eilmer: refactor cell and face geometry calculation code.
Have moved the code for computing primary-cell geometry out of the UBlock class
and into the FVCell and FVInterface classes.  Together with the reverse-lookup
from the previous commit, this refactor will allow reuse of the code when Kyle
wants to update just a few cells and faces that surround a particular vertex.