Eilmer: Bug fix when listing actions in the Gas-Solid adjacent BC.
I removed the useless pre-spatial derivatives action that
tried to compute an interface temperature. We don't really
need this at that stage, because we are going to compute
a flux directly in a subsequent step.

There was an internal bug in that the boundary interface
effect was never initialised, so it never picked up the
cells and interfaces it needed to work on. This problem
was never noticed because when one loops over (0 .. 0)
it has no effect. Since this BC had no effect, and yet
all our verification tests passed, it seemed safe to
remove it from the list of actions.