Eilmer: a couple of notation changes, to be a little more consistent.
When specifying flow state for inflow conditions, whe have changed:
old-notation  --> new notation
flowCondition     flowState
stagCondition     stagnationState

In the code we try to use the new name but, if it is not found,
we fall back to using the old name.

Also, with this commit, only a few examples have been changed.
The rest will follow.