Eilmer: Loads now write separate files at interval of dt_loads...
and there is a file called jobName-loads.times that is written
in the loads/ folder to indicate the mapping from loads_index
to simulation time.

NOTE: The loads files and -loads.times file reset from 0 on
a restart. This is because the restart index has no connection
to the loads_index. I didn't want to build a search into the
-loads.times file based on an arbitrary simulation time restart.
If this use-case arises, then we'll address the issue. For the
moment the 'caveat emptor' is that the user needs to be careful
about storing a copy of the loads files if transient results
are important AND they plan on doing restarts.