Eilmer: Added --next-loads-index command-line option.
In a recent update to the loads output, I had coded
that the loads index would start agin from 0 on
restarts. This was probably going to be an annoying
situation when using the code on queue-managed cluster
computers where restarts are frequent due to batch
run time limits.

In this commit, the user can now explicitly set the
next index for the loads output when doing a restart.
If they do NOT set anything, but are attempting a
restart, then the code will attempt to find the
final loads index from an earlier run. It will
then increment that value by 1 as the starting
value for the next loads output to be written.