kinetics/gas: Moved kinetics-related modules into kinetics package.
This commit is a rather large churn of code that is mostly
about moving the furniture. I have separated all of the pieces
that are only concerned with the kinetics of changes in the gas
state into the kinetics package.

This gives the gas and kinetics packages a cleaner division
of responsibilities:

  gas package: handle the computation of thermodynamic state
    and trasport properties

  kinetics pacakge: deal with rates of change of a gas state
    due to kinetic effects, such as chemistry and energy exchange.