Eilmer: eliminate the code for computing distances to nearest wall.
This code has been giving trouble, partly because it was trying to do a couple
of different things and partly because there was a latent bug for some years.
Since it is not being used for its original purpose but just to get
the cell-half-width to the wall for a couple of boundary conditions,
it seems simpler to just compute the distance when needed in the particular
boundary-condition actions.