eilmer: added in -- unstructured 2nd-order & first stab at VTK plotting
This adds in the ability to run the adjoint solver for 2nd order
unsturctured simulations. It was discovered that when forming the dRdX
sentivity, with high-order reconstruction effects, it is necessary to
also update the ghost cell center for perturbations of boundary vertices.
For now we will apply first order reconstruction at boundaries.

This also adds the ability to plot the adjoint variables for both the
structured, and unstructured solver.