examples/Eilmer: Fixed up cht-manufactured solution to work with new T nomenclature.
Also, the subtract_ref_soln methods will just do nothing if the user does
not supply an appropriate reference function. This de-couples the fluid
and solid domains. If one only wants to supply a reference function for
the solid domain, for example, then the implementation just ignores
a missing reference function for the fluid domain.