eilmer: refactor adjoint solver flux routine to resemble flow solver routine
The adjoint solver flux routine now operates on a list of arrays. This
allows only the interfaces that are effected by a cell, or vertex
perturbation to be recomputed, thus improving efficiency. Also,
now that the adjoint flux routine resembles the flow solver routines
it will be easier to maintain.