Eilmer: eliminate fromTable and toTable methods; they should not be needed.
They were used in the construction of initial flow fields from previous
simulation data.  Now that the Lua domain FlowState is a table and the
D-language domain accepts tables of suitable form for writing the initial
flow field data, we should be able to just use the table returned by the
FlowSolution class for a cell.