eilmer: added MLP limiter, and node-neighbour stencils for lsq-reconstruction
Following recommendations outlined by the recent paper introducing VULCAN's
unstructured solver (White et al 2017) the following has been implemented to increase the
robustness of the unstructured solver:

+ MLP limiter -- the implementation of the MLP limiter follows that of the VULCAN
code, i.e. the MLP framework is used in conjunction with the Van Leer limiting function.
Implementation of the MLP limiter requires data storage at the vertices, currently the
LSQInterpGradients object is used to store the data, however only half of the contents
of the LSQInterpGradeints object are needed at the vertices. Profiling of the code will
be undertaken in the future to determine whether this needs to be refactored for excessive
memory consumption.

+ node neighbour stencil -- for the inviscid cell-averaged gradients, it has been suggested
that a larger stencil be used. To be consistent with the implementation of the MLP limiter, the
extended stencil uses all cells attached to the nodes that make up the parent cell. This stencil
is used when the use_extended_stencil flag is switched to true in the lua input script. To consistently
apply this stencil across the domain for multi-block simulations requires more cells than currently communicated
to neighbouring blocks, so the current implementation reverts back to the compact stencil for boundary cells.