gas: Changed thermo coefficients in two-temperature air model.
The thermo coefficients for the enthalpy curves  provided in the Gupta et al (1990)
report are dodgy near the 300 K: they do not agree with a simple Cp=const extrapolation
from 298.15 K (reference temperature) up to 300.0 K.

The thermo coefficients in the Gnoffo et al (1989) report give more precision and the
integration constant for enthalpy is better behaved near 300.0 K. So I have put these
coefficients into the model instead. There is still some very small discrepancy at 300 K,
but not enough to worry the iterative procedure to compute temperature from energy.
Even so, I have added a linear blend for enthalpy in the region between 298.15 K(=T_REF)
and 300.0 K so that the reference enthalpy value is honoured and so that there is no
discontinuity when switching to the polynomials at 300.0 K.