kinetics/gas: 2-T air mode now gives reasonable results compared to experiment.
In the gas module, I needed to ensure that the heats of formation were consistent
with the change to Gnoffo's coefficients.

In the kinetics module, I needed to make a call to rhoT at the end
of the chemistry update to account for the shift in internal
energy distribution when mass fractions change.

In fvcell, I added an extra piece of information about the cell's
ID in the case of a failed thermochemical update.