Eilmer: Make FullFaceCopy more like MappedCellCopy.
This GhostCellEffect class now sets up lists of ghost_cells and mapped_cells
shortly after the grids are constructed and then uses this lists when the
boundary condition effects are applied.

A result of this refactor is that the encoded mapping of the ghost-cells to
other cells is moved from the sblock module into the ghost_cell_effect module.
The bulk of the changed lines in this commit is associated with this move.

Breaking news: Although most of the test cases have been running successfully
with this latest change set, I see that the block-marching case of the channel
with bump has just failed.