Eilmer: added globalFluidBlocks array to hold all blocks in MPI-parallel.
Those blocks in the current MPI-task or process will be complete with their
internal data (grid and flow) while blocks that are handled by other MPI-tasks
will have their internal data incomplete.  There will be configuration data
but not the actual grid and flow data.

In the shared-memory flavour of the code, the global- and local-FluidBlock
arrays will hold references to the same (full set of) block objects.