eilmer: makefile now finds MPI libraries when not set as part of system default
For any non-standard installs of MPI, we need to tell the linker where to find
the MPI libraries at link time. I needed to do this on my CentOS system and
when compiling on tinaroo. I figured this problem might be more widespread
on other cluster systems. I've added a little search in the makefile that
will try to locate the appropriate MPI lib directories.

With this addition, this obviates the need for setting LIBRARY_PATH. I have removed
that part of the INSTALL instructions.