nm: ILUp pre-conditioner now uses sparse matrix.
Unlike the ILU0 pre-conditioner, the ILUp pre-conditioner does not
use the same zero-pattern as the input matrix. As a consequence, the initial
implementation of the ILUp pre-conditioner required a large amount of memory
to store the original sparse input matrix, in addition to the two extra matrices
required for the algorithm (the pre-conditioner matrix, and the levels matrix).

Now that we allow non-zero values to be assigned to initially zero entries in the
SMatrix class, we can use a sparse matrix for the pre-conditioner matrix in the
ILUp method, which should hopefully reduce the memory overhead for this pre-conditioner.

Note however that some stored entries will be zero, this is a result of entries
starting as non-zero in the original matrix, and then becoming zero during the factorisation.