gas: Increase robustness when bracketing the temperature value.
When trying to find a good bracket for the iterative temperature
solve, I had a rather aggressive choice that relied on the
given temperature value being a good guess. When there are large
changes, the old temperature value is not necessarily a good guess.

So now there is a fall-back position: when the aggressive search fails to
bracket the value, we then use a very conservative selection of temperature
brackets. We start searching between 20 K and 100,000 K. If that search
fails, then I suggest looking elsewhere than the thermochemistry to blame.