eilmer/gas: Use species names of form 'NO+' and 'e-' in user tables.
The user can now use symbolic names for setting the species in
the input tables. Note that the ions and the electron will need
to be protected with brackets when used in a table so that '+'
and '-' aren't interpreted as operators. An example is:

inflow = FlowState:new{p=p_inf, T=T_inf, velx=u_inf,
                       massf={N2=0.67878, O2=0.04483, N=0.00015,
                              O=0.25403,NO=0.02221,['NO+']=0.0}}