eilmer: changed filename adjoint_solver.d to optcore.d
To be more precise I have changed the name of the optimisation back-bone
to optcore.d (to be consistent with simcore.d -- the back-bone of the flow
solver).

This is more precise since the adjoint solver only makes up part of the
gradient computation. In optcore.d we house the adjoint solver, the
gradient calculator (used in combination with the adjoint solver output), and
a finite-difference solver (for verifying adjoint gradients).