eilmer: added in function for forming only diagonal of a 1st order Jacobian.
For the steady-state solver preconditioner we would like just the diagonal
of the 1st order Jacobian. I have added a function at the end of the
shape sensitivty calculator that borrows some of the methods from forming
the full Jacobian. Note that there is one extra step for forming the
steady-state preconditioner, that is the multiplication of the Jacobian
by the transform matrix (to convert from primitive variable form to
conservative variable form).