eilmer: Fixed bug in constructing lower order Jacobians.
I was setting the order of the interpolation to the lower order in
the Jacobian construction code, housed in the shape sensitivity calculator.
However, I wasn't setting the config interpolation order to the lower
order, thus the flux calculator was still using higher order interpolation,
which was ultimately causing NaN's to be generated for some simulations due to
the inconsistency

This update has fixed the issue. Also, I have broken up the steady-state solver
preconditioner into two stages: an initialisation stage, and a construction
phase.