Eilmer4: Commit Mabey's turbulent flat plate examples
Commit Mabey's turbulent flat plate examples. There are two sub-folders
in the flat-plate-turbulent-mabey folder; one contains a setup script
written for a non-wall-function (solve to the wall) k-omega simulation,
and the other for a simulation with wall function (with a y+ set to
approximately 100 -> a good value to use).

Note that both these scripts use the InFlowBC_StaticProfile to import
a inflow boundary layer profile on the South & West boundaries.

Also, many thanks to my MECH4501 student, Dimir Pot, for his continuing
efforts in getting the wall function capability validated.