eilmer: Infrastructure for diagonal preconditioner in place for e4sss
The code still produces the same results when "use_preconditioner=false".
That's a good thing.

With the preconditioner on, there are problems. However, I wanted
to get the code committed after doing the refactor when it's in a
state that the non-preconditioned code is doing the same job as
it did prior to the refactor.