eilmer: the convective flux and unstructured limiters now take the gtl as an argument.
Previously we were assuming gtl=0 for the convective_flux_phase0(), convective_flux_phase1(), and
the unstructured limiters. This was fine for the flow solver, however, in the shape sensitivity calculator we
perturb the mesh and store it in the grid-time-levels, then we perform a residual computation on this perturbed
mesh to compute the sensitivity of the residual to perturbations in the mesh/design variables.

In this update, the functions mentioned above now take the gtl as an argument, so they may compute the flux on the correct
gtl. The default for the functions are set to work on gtl=0, so all current instances of these functions in
the code keep their original functionality.