eilmer: fixed a couple of bugs in e4opt & e4ssc.
e4opt:
I was interpreting the ASV value in the DAKOTA params.in file incorrectly. Note that
a value of 1 means DAKOTA is expecting the objective function be evaluated. A value of
2 means thats DAKOTA is expecting the gradients to the objective function be evaluated.
A value of 3 means that DAKOTA is expecting both the objective function, and the
gradients be evaluated.

e4ssc:
I was writing out the objective function evaluation for every execution of e4ssc. This
confused DAKOTA. Now we only write out the specific information DAKOTA has queried.