eilmer: steady-state preconditioner now can use the block diagonal of a first order Jacobian.
Each block making up the diagonal of a first order Jacobian is stored in the cell object as a
n x n matrix, where n is the number of primitive or conservative variables.

NB. the transformation from primitive to conservative variable form currently assumes
a 2D simulation. Further extension is required for 3D, and turbulent simulations.