eilmer: we may now run 1st order viscous simulation using the unstructured solver.
We had hidden the initialisation of the weighted leaste squares workspace, used in
computing the spatial derivatives, behind an if (interpolation order > 1) condition.
I have shuffled around the if statement to now only act on the inviscid reconstruction
least squares workspace, which allows us to intialise the weighted least squares workspace
even for first order simulations.