eilmer: logic to select consistency of spatial deriv calculation updated.
I asked Peter to review the logic and we've arrived at the new
selections for the spatial derivative location. The motivation
behind this is that we'd like the default of divergence theorem
at vertices for 2D calculations on structured grids. This gives
the best emulation of eilmer3 behvaiour and leads to fewer surprises
on structured grids.

However, that default is not appropriate in 3D nor on unstructured
grids. In which case, this consistency check will pick that up
and make the appropriate changes. It informs the user if it
is making changes.