eilmer: implemented correct adjoint boundary conditions for finite-difference linearised adjoint solver.
Initially the boundary conditions were being applied for every cell perturbation. This gave reasonable gradient
verification results. However, this is fundamentally not the correct method for forming the adjoint boundary conditions
by finite-differences. If we apply the boundary conditions after a boundary cell value is perturbed,
we are actually computing the effects of the boundary cells perturbation and the ghost cell perturbation all at once, which does not compute the true
partial derivative. Instead one should compute the effects from the internal perturbation, and the ghost cell perturbation separately
and sum the contributions.

In this update we follow the method outlined in the literature for forming the flow Jacobian using finite-differences
with correct adjoint boundary conditions. The particular paper to reference for this implementation is "The Use of Finite-Differenced
Jacobians for Solving the Euler Equations and for Evaluating Sensitivity Derivatives", Rizk, 1994.

The new boundary condition application has been applied to the adjoint transposed Jacobian, and the steady-state solver preconditioner.

NB. Currently the code isn't written in the most maintainable or easy to follow format. A refactoring of the code will follow. I am pushing
the current functioning state so as to not delay work on the sss-preconditioner.