NM: Begin generalizing the numerical methods package to use complex numbers.
Kyle has been experimenting with "complexification" of the flow solvers
in a fork of the dgd code.  It is now time to port this complex-numbers
capable code back into the main repository.

We are starting numerical methods because they have the least dependencies
and because the basic complex number class is housed there.

This commit includes Kyle's adaption of the Complex class from D's standard
library.  He needed to add a few methods to the existing class.

This commit also introduces the nm package so that we have the adapted Complex
class in a distinct namespace.