NM: by default, omit the body of the complex module.  Sounds weird but...
It seems that something in the complex module breaks the DMD compier optimization
for versions of the DMD compiler more recent than v2.76.1.

Since the default build of the code includes optimization but most users of the
code do not care about the complex_numbers flavour of the build, we just hide
the troublesome code unless it is requested explicitly with the compiler option
-version(complex_numbers).

Note that all recent versions of the DMD compiler happily build the complex code
without optimization and that the ldmd2 compiler (v1.10-beta) builds the complex
code with optimization requested, so we have good options to press on with
development.