Eilmer: Provide a close() method for the lua-wrapped FlowSolution.
The FlowSolution objects are typically held onto in memory to avoid
the D garbage collector from throwing away the object before Lua
is finished with it. However, for large flow solutions, or opening
many of them, the memory burden of holding onto all of those objects
can get large.

This commit provides a "close" method for the lua-wrapped FlowSolution
object. Users should call it when they are finished with a FlowSolution
object to free up the memory. The close method will release most of the
internally stored memory by throwing away the flowBlocks and gridBlocks.
Note the object itself still persists in the D flowSolutionStore until
the program exits. In other words, a FlowSolution is reduced to a shell
after calling close.