NM: make the Matrix class of bare-bones linear algebra into a template class.
The type of the data elements held by the Matrix is now parameterized.
The flow code has been updated to instantiate Matrix objects with
double elements and the Arc3 class instantiates the Matrix objects
with number elements.

There is a bug in the evaluation of the centre of the arc, I think.