geom: modified conditional statement when deciding if to flush small components to zero in normalize() method.
Originally we were just checking the magnitude of the real part of the Vector3 components to decide whether to set that
component to 0, however, for the complex numbers version this was erasing non-zero imaginary components. The new modification
checks if both the real and imaginary parts are sufficiently small to set them to 0. In addition, we now only set the entry to 0
if the magnitude is below 1.0e-30 instead of 1.0e-14.