NM: Parameterize the data type for mose numerical methods.
The templates now take the data type as a parameter so the code depending on
these template functions now needs to be updated.  Thus this commit is NOT for
general use but is intended for review.  I need a fresh set of eyes of the code
to see what I've done wrong with the function minimizer from nelmin.