Geometry: now have complex-number derivatives working for Line and Arc.
There are tests for each of these two classes that set up a finite-difference
calculation of the derivative of mid-point position with respect to a parameter,
alpha, that was used to define the paths.
The perturbation is along the imaginary axis for each of the complex components
of the defining Vector3 points.