eilmer: fixed 2 major bugs in shape sensitivty calculator.
1. We were not updating the cell viscosity and thermal conductivity when perturbing the primitive
variables for construction of the flow Jacobian. This has been amended.

2. We were originally storing the perturbed grid in the gtl=1 (grid-time-level) when computing the
residual vector sensitivity. The issue with this approach is that the references to the cloud points
for the spatial gradients are a combination of interfaces and cell positions. The interface positions
do not use the gtl array system, and thus when the geometry was being updated on a perturbed mesh, these values
were being correctly stored and used. However, the cell positions do use the gtl array system. The problem is that the
reference in the cloud array is pointing to the gtl=0 position, so the updated cell position wasn't being used
in the residual vector sensitivity evaluation (which was acting on gtl=1). We now compute the residual (via evalRHS)
using the gtl=0 position.