kinetics: Add the skeleton of kinetics for a gas with a mixture of pseudo-species.
This is a work-in-progress commit. The basic layout of an implementation
for state-specific kinetics is in place. There is some more work to go
with the help of Pierre to fill in the details.

Also, I've disabled the pseudo_species_gas model during this development.
The internal structure of that implementation will change to support
the kinetics implementation.