eilmer: ability to enforce 1st order reconstruction in cells capturing a shock for the unstructured solver.
In the unstructured code, spurious errors can result due to the higher-order reconstruction methods
in the presence of strong shock waves. To increase robustness, we may now enforce cells
in the shock to apply the first order reconstruction. This trick is taken from Candler and his
work on unstructured hypersonic flow codes.

To set first order reconstruction in the cells capturing shocks, apply:

       config.suppress_reconstruction_at_captured_shocks = true

remember to also set the compression tolerance used in the shock detector to a suitable value!