This commit includes the first functional model for an electronically specific state to state solver.
The primary inclusion in this commit is the functional electronic_state_solver.d.
While the model is functional, it has not yet been implemented in the cfd solver
until complex numbers are included in the package.
Appropriate changes were made to the associated kinetics and gas packages to implement the
multiple electronic gas states for nitrogen and oxygen.
Edits were made to include these packages in the appropriate makefile and package test,
although no test scenario has yet been included.