eilmer: added Hanel's flux calculator.
As suggested in Wada & Lious 1997 SIAM paper,
   "An accurate and robust flux splitting scheme for shock and contact discontinuities"
I have made available the high-dissipation flux calculator they attribute to Hanel, and thus
is referred to as Hanel's flux calculator. The reason for this addition is that it is suggested
to be a good match with AUSMDV for removing the carbuncle pheonemon. The adaptive blending will be
added in a later revision.

NB. compared to our other two high-dissipation flux calculators, EFM & HLLE, Hanel's flux calculator
appears to have a superior convergence rate - at least for the Mach 6 through to Mach 16 flow over
a cylinder suite of tests I ran.