Gas: Part way through the process of making the gas models mostly @nogc.
*** This commit is for review only; do not use it for your work. ***

Some of the changes have been to the construction of exception messages,
where -dip1008 is used to allow throw new Exception(msg) to be @nogc.
More complete messages are constructed only in debug blocks.

Other changes have been to replace some of Rowan's thermo database dictionaries
in the two_temperature_air model with simple arrays.

Other quick-and-dirty changes are to implement the method calls for some
rarely-used gas models in debug blocks and throw expections if they are used
in a non-debug build of the code.  Clearly, this all needs fixing...