Eilmer: use a return flag from decode_conserved and check it...
in a couple of places where it would be a surprise to find cells with bad data.

In the update functions we will continue to ignore each return because we will
count the number of cells with bad data immediately.

Simplified the name of the flag that each cell carries.  It may be more than
just the thermo data that is bad.