eilmer: re-implemented scaling for the steady-state solver.
We may now scale the equations and unknowns of the linear system arising
in the steady-state solver. This scaling is also now compatible with
preconditioning. Scaling makes the solver more robust, especially for
turbulent simulations, where the flow variables can have quite different
orders of magnitude.

The implementation of the scaling is based on algorithm 3.2 from:

    Matrix-free methods for stiff systems of ODES,
    Brown and Hindmarsh (1986)

in this implemenation, the column and row (equations and unknowns) scaling
matrices are the same diagonal matrix, this is also suggested in the more recent paper:

    A Jacobian-free Newton-Krylov algorithm for comrpessible turbulent fluid flows,
    Chisholm and Zingg (2009)

To enable scaling, set:

   config.use_scaling = true