eilmer: modified the k-omega boundary interface effect.
When using turbulent zones we would like to not use the k-omega model
to set the boundary interface tke and omega values for interfaces outside
of the turbulent zone, since this causes trace amounts of tke and omega
to bleed into the cells which are marked as *not* in the turbulent zone
(this showed erratic behaviour for some turbulent simulations using
the steady-state solver).

To remedy this, I have modified the boundary condition to only apply
the k-omega boundary values if the internal neighbour cell is in the
turbulent zone, else we just copy the internal cells tke and omega
values to the interface.

The new modification showed markedly better convergence.