kinetics: Electronically specific gas updated to include dissociation reactions for N2 and O2
This commit includes chemical reaction rates taken from Gupta et al for the dissociation
and recombination of N2 and O2. This is only loosely coupled to the other electronic states.
Currently, the reaction seems overly forward biased. Need to confirm behaviour.

The kinetics solver for the electronic states was adjusted to primarily solve each step
with a second order backwards differentiation method. The appropriate numerical methods
in bdfLU were cleaned up to be implemented in the same manner as the first order method.
This significantly improves numerical stability.