eilmer: enforce positive values of tke and omega in k-omega turbulence model.
For stability of the steady-state solver it is recommended by some sources to
prevent the tke and omega variables from going negative. Some authors clip
the turbulence variables by setting them to 0.0; as has been done for the
Spalart-Allmaras model. To prevent a division by 0.0, we instead set the tke
and omega variables to very small non-zero values. A value of 1.0e-10 has
proven to be robust from initial testing.