eilmer: made default spatial derivs least-squares boundary stencil more robust compact stencil.
It appears from numerical experiments that the k-omega turbulence model produces non-physical artefacts
at block-block boundaries when using the symmetric spatial derivatives least-squares stencils developed in
Verification of the Least-Squares Procedure within an Unstructured-Grid Flow Solver (Damm et al, 2016).

To remedy this problem, the default stencil has been changed to a more robust compact stencil that consists of
the ghost cell and internal cell attached to the boundary interface, in addition to all the interfaces that make
up the internal cell. This removed the previously mentioned artefacts.

Currently the symmetric stencil cannot be selected, however a user configurable variable to allow
selection of the symmetric boundary stencil for use in special cases (for example MMS verification) will be
added in a later commit.