kinetics: Removed energy relaxation between modes in the electronically specific kinetics.
Energy relaxation between temperature modes proved to be too complicated for the purpose of
this initial electronic kinetics model. The Appleton and Bray model was implemented, and
seemed to work, but energy transfer was too slow. This is likely due to the implementation
being limited to the translational-electronic relaxation, and did not account for radiative
or vibrational energy transfer. It's recommended that analysis into the significantly populating
rates be done to select modes to model.

Lua wrapper was also edited to allow for dtSuggest to be passed to the kinetics solver.