eilmer: scale omega (from k-omega turbulence model) residual by 1/1000 for dt evaluation.
The omega variable in the k-omega turbulence model appears to be extremely stiff, especially so on
grids with y+ <= 1. This is suspected to be related to the omega boundary condition; theoretically
omega goes to infinity at a boundary - of course this is truncated in reality, however very large
omega values still exist at wall boundaries. This stiffness causes the omega variable residual to
fluctuate by a large amount when using the implicit solver (a few order of magnitude in some circumstances),
which dominates the global relative residual, which the time-step growth algorithm is based on. As a means to
stabilize the time-step growth the omega residual is divided by 1000 before the global residual is evaluated.
This appears to be sufficient enough to stabilize the time-step growth in a handful of test problems.