eilmer: fixed multi-block bug in shape sensitivity calculator.
For multi-block simulations, when we re-compute the spatial derivatives
stencil values and primary geometric cell properties (which the former
is a function of) after a perturbation of the grid, we must transfer
ghost cell data in between the two steps. The reason is that the updating
of cell primary geometric properties fills the ghost cell positions with
extrapolated estimates not the actual neighbour cells new position. So,
we must first compute the new cell geometric cell properties, then transfer
ghost cell data, then update the spatial derivatives stencil values. With this
bug fix, viscous adjoint simulations achieve machine precision accuracy
compared to the direct complex-step method.