Eilmer: changed to using OutFlowBC_SimpleFlux for OutFlowBC_Simple
This leads to some small differences in the Giordano cylinder steps,
and some not-so-small changes for the t4m4 block-marching nozzle calculation
(hence the explicit use of OutFlowBC_SimpleExtrapolate for that case).
The other cases of explicitly calling OutFlowBC_SimpleExtrapolate are where
we want to specify the xOrder of extrapolation for the manufactured solutions.