Eilmer: Allow filename as initialState in FluidBlock
When supplying a filename as the initial state, it is
assumed that the user is supplying another eilmer flow
field file. No checks are done. To be a valid a initial
state, the supplied flow field file must be:

 + of the same type as requested in this simulation, eg.
   gzippedtext, rawbinary, etc.
 + have exactly the same cell topology

This option is considered for use by advanced users.
It is available to facilitate the transfer of flow
field information between grids of identical topology
but not necessarily identical geometric location.

The use case for this option is warm-starting
optimisation calculations. In these cases, using
an old flow solution is a desirable starting point,
however the grid positions will have shifted due
to the optimiser morphing the design geometry.