kinetics: Simplify responsibility for chemistry: now only update mass fractions.
This change is to make the chemistry updater more easily adapted for use
in large, more complex thermochemical update. The chemistry updater made
the assumption that it would apply the thermodynamic constraint of
constant energy, constant mass. In other words, a fixed volume reactor.

However, it's better to limit the responsibility of the chemistry
updater to only change mass fractions over some small timestep.
Then the caller/user of the chemistry updater can apply his/her own
thermodynamic constraint after the update service returns.

So restricting what the chemistry updater does actually makes it more useful.
This change is in preparation for including the chemistry updater as
a sub-service in the 2-T air thermochemical update class.