Eilmer: Check that we do not ask for features that have been omitted.
Now that version(xxxx) can be used to include or omit certain features,
we should check that the simulation does not require any particular capability
that happens to have been omitted.

It is not always obvious what capabilities are required.  For example,
asking for a perfect-gas model of air actually uses a multi-species
gas model even though the user thinks of it as a single-species model.