eilmer: adjoint pseudo time-stepping bug fix.
The precondition matrix was accounting for the 1/dt term incorrectly. This
revision fixes the bug. Because it is costly to perform the ILU decomposition
of the precondition matrix each time the CFL is ramped, we chose to use a
fixed CFL (for now) - this approach is also used by Nielsen in his thesis.