Geometry: Restart work on porting Heather's paver for 2D grids.
This present commit is a bit of a cheat in that it generates a mesh
of triangles, all with a common point in the middle of the domain.
Although, this is useless for flow simulation, it confirms that the basic
data plumbing does indeed connect the new paver to the USGrid class.