kinetics: Structured the electronically specific kinetics to be based off the new two temperature kinetics.
Most of the structure of the two temperature kinetics model was copied to the electronically specific kinetics.
Some changes were made to specify macro species kinetics which will rely on generating a macro gas state.
This macro gas state will be privately defined in the kinetics file.

Adjustments were also made to the electronically specific gas file, with regard to how it performed thermo updates.
Previously it relied on being passed a macro gas state, before perfoming the same thermo updates as the two
temperature model.
Instead, it now is passed the full gas state, and updates a stored macro species gas state for thermo updates.