kinetics: Restructured the methods and classes for much of the electronic state solver to be more in line with existing eilmer files.
A lot of the pre-existing electronically specific model was written as a stand-alone file, and massaged in to the eilmer kinetics
module to try to run it as a kinetics model. This was done prematurely at the time, and the code structure was not suitable for
continued maintenance or expansion. A lot of this, was mostly due to my inexperience in the D language at the time of writing.

The current electronically specific method tries to emulate the style of the chemistry update file. The electronically specific
interactions are of course similar reactions to those between different chemical species. The numerical steps are still solved
with an implicit method written in the nm/bdfLU.d file, as they are more tightly coupled than most other chemical reactions.

Thanks to the code restructure, the method by which data is able to be input has also changed. Similar reaction mechanisms
to chemical reactions are now described for interactions between electronic species (sample-input/ESK-N.lua for example).
This allowed for much simpler initialisation of reaction objects in the kinetics file, enabling electronic reactions to
be easily enabled or disabled depending on flow conditions. The development of electronic_update class objects also
allows for the simple expansion of the model to include many more electronically specific levels within other chemical species.

The original electronically specific state solver file was removed, along with the older sample-input files for reaction
rate fitting parameters.