Eilmer: history-cell files are now versioned.
History-cell files now have an integer suffix counting from 0.
Each time a simulation is run and the history files are initialized,
a check is made for history files from previous simulations.
If such files are found, the new files will have their suffix incremented
such that the new data is not appended to the old files.

To assemble a sequence of files into a single history file, just concatenate
the individual files.  If you cannot tolerate the comment line appearing
periodically through the combined file, you will need to remove that line
from each of the individual files as they are being concatenated.