Gas: remove the species concentration array introduced in the previous commit
because it introduces a circular dependency between mass-fractions, density and
concentrations if we try to work from concentrations in some of the reacting
gas models.  We revert to using only density and mass-fractions when sending
data from the flow solver to the thermochemical functions.

This commit also includes:
(1) some species symbols in the Argon2T model and it adds
the test for this gas model to the list in gas-package-test script and
(2) the introduction of the charge-balance method to the GasModel class.