Gas: Rework Daniel and Rory's argon gas model to be @nogc.
This rework allows the model to be used in an optimised build of the code.
To do this, the use of Matrix class is avoided and fixed-length arrays are
used instead.  This uglifies the code a bit, particularly when explicit loops
have to be used to be compatible with the complex numbers build.

Also, note that this commit has broken the Backward_Euler integration.
To end on a positive note, note that the RK4 and Forward_Euler integrations
do pass the unit test.