eilmer: interface flowstate for viscous flux now computed from cell-centered averages.
Originally we were averaging the reconstructed flowstate at either side of an
interface during the convective flux update, for later use in the viscous flux
update. This resulted in the stencil of influenced cells being larger for the
viscous fluxes compared to the inviscid fluxes.

This update now averages the neighbouring cell-centered values either side of
the interface. Currently it is only implemented for the unstructured solver.