bc: remove reverse_tangential_velocity from internal_copy_then_reflect.d
It has been a TODO to remove the reverse_tangential_velocity method. This
was added back at the start of the unstructured code developement as a trial
fix for some oddities we were getting in boundary layer flows. The current
version of the code appears insensitive to this piece of code now, and thus
it is time to remove the ugly hack.