eilmer: added i_bndry to fvinterface object.
This is in preparation for extending the boundary condition handling
to accept the application of the boundary conditions on a per interface basis.
We need the index in bc.faces to obtain the correct outsign stored in
bc.outsigns for a particular interface - this is what we will now store
in i_bndry.

TODO: extend to structured solver, if necessary.