Eilmer: infoFluidBlock now returns the structured-block corner positions.
In 2D, these positions are fields p00, p10, p11 and p01 in the returned table
of information.  The values of these fields are tables with fields x, y and z
attached to double values.  In 3D, the names are p000, p100, etc.

The actual coordinates should be current for all structured-grid blocks,
even for those blocks assigned to another task in a parallel MPI simulation.