eilmer: Allow --prep mode to only work on a restricted range of blocks.
When using prep, the user may specify a restricted range of blocks.
Blocks to work on are separatde by a comma (","), and a range
may be given as start..<end which goes from index 'start'
up to but no including 'end'.

As an example:

> e4shared --job=bump --prep --only-blocks"0,1,17..<21"

will builf flow files for blocks: [0, 1, 17, 18, 19, 20]