Eilmer: Broadcast the master userPad and have it appear in all Lua interpreters.
The userPad array may be read and written to in all interpreters but,
if your Lua functions alter the values, only calls to atTimestepStart,
atTimeStepEnd and atWriteToFile will preserve the newly assigned values.

Also, for MPI jobs, the userPad values for the master task (rank 0) will be
broadcast to all other MPI tasks and then propagated through all of the Lua
interpreters in each task.  Setting the values in tasks other than the master
will likely lead to disappointment.