Eilmer: simplify the code for the venkat limiter and optionally suppress it at edges.
For no-ghost-cell boundaries, we need to suppress the limiter to get good
extrapolation, so only apply the limiter at edges if extreme_clipping is also set.

Also, relax the tolerance for surface pressure in the cone20 unstructured-grid
cases to allow WallBC_WithSlip1 boundary conditions pass the Tcl tests.

At this point, all automated tests should be passing for WithSlip0 and WithSlip1
boundary conditions.