Eilmer: expansion tube demo with diaphragm, MPI simulation.
This makes use of the recently added ExchangeBC_FullFacePlusUDF boundary
condition to allow the conditional exchange of boundary data for a distributed-
memory simulation.  We no longer need to keep all diaphragm-connected blocks
on the master mpi task, however, we still need to keep the coordinating blocks
on the master.