Update shear_stress_relative_limit code.
(a) Increase the default value to 1.0.  Kyle reported that the 0.25 value
interfered with the double-cone viscous simulations.
It is not clear what is a good value; some thinking and testing is needed.
(b) Introduce config.apply_shear_stress_relative_limit with default value of
false so that you have to deliberately set it true to have the limiting active.