Eilmer: We now have a monitor program.
The program e4monitor can be started in the simulation directory and
will check the content of the config/<job>-progress.txt file for changes.
The simulation program will write its current time step and, on completion,
the word "done" to that file.  This will be an indicator of progress.