eilmer: make proper use of the mapped-cells for transferring viscous gradients via mpi.
The unstructured solver requires spatial gradients (sometimes referred to as viscous gradients)
along block-block boundaries to be communicated.
This had been implemented into the unstructured steady-state solver framework late last year (2018),
but was not extended to the explicit solver.
This update now makes proper use of the mapped-cells to communicate the viscous gradients for
(viscous) domain decomposed simulations using the shared-memory and MPI flavours of the unstructured grid explicit solver.