Eilmer: Added writing of residuals to own subdirectory.
The residuals are written in Eilmer-native format.
They are presently written only when running the steady-state
solver.

The implementation is incomplete. It only considers the residuals
of four conservation equations presently: mass, x-/y-momentum, and
total energy. I'll need to add other residuals when the
steady-state solver is more widely used.