Add a Ruby script to collate data for building x,t-diagrams.
The script works through the solution files for a job, taking a slice from
each solution instance and saving selected data to a text file that can be read
by GNUPlot.  The actual x,t-diagram can be drawn from this collated data.

Also, add an example of use to the expansion-tube-demo/mpi/ case.