Eilmer: revisit the simulation of a boundary-layer growing along a cylinder.
This is a pair of low-resolution, axisymmetric, simulations that test
the capture of a viscous boundary layer.  The peak-temperature point in the
boundary layer profile, toward the end of the cylinder, is tested against
some numbers from David Pruett's spectral solution.

Now that they are added to the set of automated tests, it may be of interest
to know that the run times are about 15 minutes each (on a 4-core workstation).