Make the gasdyn gasflow functions available in the CFFI gas module.
So far, the first of the functions (shock_ideal) is made available via
the GasFlow class.

We had to do a bit of fiddling with the makefiles to get the gasflow D module
into the loadable library without the Lua wrapping functions.

Along the way, we moved the billig.lua module into src/lib/.