Gas: improve the behaviour of the Lua wrapper for the CEAGas model.
Recovering the ceaSavedData from the Lua table into a newly-minted GasState
object is needed for some of the service functions that expect to find data
in place.
Also, in a couple of places, we needed to call update_thermo_from_pT to get
a full set of data in place.