eilmer: created a new Block class. FluidBlock & SolidBlock now are derived classes from the base Block class.
This revision also removes the globalFluidBlock[] and globalSolidBlock[] arrays, replacing them with a single globalBlocks[] array.
The new global array will eventually host the collection of both fluid and solid blocks, for now it only acts as a replacement for
the globalFluidBlock[] array.
This is in preparation for the solid solver MPI implementation.