eilmer: solid blocks are now stored alongside fluidblocks in globalBlocks.
Note that we assume all the fluidblocks are in the first 0 to nFluidBlocks
entries of globalBlocks[] and all solidblocks take up the remaining nSolidBlocks
entries.

Currently the solid block solver still only functions in shared memory parallel,
these modifications are simply cosmetic in preparation for the MPI implementation.

Note, this revision also removes the dedicatedSolidConfig[] array, we now use
dedicatedConfig[] array for both fluid and solid blocks.