Gas: added TwoTemperatureArgonPlusIdealGas model.
This is a composite gas model that combines the IdealGas model and the
TwoTemperatureReactingArgon gas model, although most of it works if
only one or the other gases is present.  Mixtures with substantial
mass-fractions of both gases are not well handled.

Also, fixed a bug in the testing of the TwoTemperatureReactingArgon
unit test and a bug in the UniformLUTPlusIdealGas model.