Eilmer: allow several group names when writing loads files.
This is a simplification of Ingo's suggestion for including names of
several groups within the config.boundary_group_for_loads string.
Group names may appear within the string with any separators (or none)
so you will have to be careful to have distinct group names that can
be used to form the file names.

Unfortunately, the name of the string now a bit misleading.