Eilmer: added Lua function mpiDistributeFBArray
This function will distribute the fluid blocks within an FBArray such that
all blocks with equal index ib will be assigned to the same MPI task.
The jb and kb indices are distributed evenly across the MPI tasks.

This is in preparation for running the block-marching calculations with
the MPI flavour of the code.