solid: added ghost cells along block-block connections.
The solid solver now utilises ghost cells along block-block boundary
connections in the same way the gas solver utilises ghost cells. Currently
some of the mechanics are in place for MPI transmission of the ghost
cell data, however, currently only the shared-memory version is
operational.

This commit also changes the spatial gradient locations from vertices
to cell-centers. This is to assist in retaining spatially second-order
solutions on domain decomposed grids.

The CHT-MMS automated test case has also been modified to allow for the changes
to the spatial gradient numerics.