examples/eilmer: Updated sphere-nonaka to reflect changes in 2-T air input.
The gas model for a 2-T air model now needs to be built using prep-gas
since it grabs some data from the species database. In this example,
one would do:

  > prep-gas gas.inp air-5sp-2T.lua

Also, DISABLED shock-fitting in this commit. The shock-fitting mode
is presently a little buggy with this case, so I've gone back to
use of a static grid.