solid: solid solver now operational with MPI.
For now this should be considered an advanced feature since it places
some onerous on the user to ensure it runs correctly. The main thing to note
is that adjacent gas/solid blocks must be assigned to be in the same MPI
task by the user in the lua input script by filling out the preassign list.

The MPI version of the solver has been tested using the CHT MMS
verification case.