nm: Change STMatrix internal storage to an associative array.
The only purpose for this sparse triplet matrix is as a staging
construct for initialising a Compressed-Row-Storage matrix.
It will be easier to do that job if the internal storage is as
an associative array. We will delete entries as we've processed
them so that at the end of intialisation, the sparse triplet
matrix has no content. This is fine, and avoids holding two entire
copies of the matrix in memory.