L1d: ported the input script preprocessor from L1d3.
Enough of the preprocessor is ported to use the new gas models.
It can read the sod.py input script example and write the old-format
parameter file without complaints.

It is not yet ready for use because it needs to write the new JSON
format parameter file, with new end-condition objects, initialize
the tube and slug data and write appropriate files.