Gas makefile: small tweak to add -fPIC, just to be sure.
It did not seem to change the state of play with the shared library.
The DMD64 build of the shared library works.
The ldmd2/ldc2 build of the shared library happens without complaint
but the library does not function correctly, with some dodgy numbers
being passed into the D functions via the CFFI interfaces
of Python and Ruby.