geom: Major rewrite to kdtree to fix a bug that caused incorrect wall distances to be computer in 3D
Basically the kdtree example I copied used a quickselect algorithm to partition
the space at each step. Unknown to me this assumed that there are no duplicate
elements in the set being sorted. Since the 3D wall examples have lines of cells
that all have the same z coordinate, this caused the partition step to terminate
before it had properly sorted the points, and the kdtree was incorrectly built.
As a result ~3% of the time it would return the wrong face, and the individual
cells could have incorrect wall distances.