eilmer: steady-state solver now has linear cfl growth option.
The default method for growing the CFL during a steady-state simulation
is based on the global residual decrease/increase. This is can be too
conservative. This commit adds in the first hack at a linear growth option.
The user may control the CFL scheduling by setting the linear growth rate
and the max CFL.

Example .lua input script (added in the steady-state solver options):

   residual_based_cfl_scheduling = false,
   cfl_max = 50000.0,
   cfl_growth_rate = 2,