kinetics: Add a test on the species order in the chemistry input files.
Since the gas model and chemistry input are separate, there was a possibility
that the species listed in each were incompatible. Previously, there was an
assumption that the chemistry file was generated at the same time as the gas
model so the risk of this incompatibility issue arising was small.

In this commit, the species are now also embedded in the chemistry file.
At runtime, the gas model and chemistry file are checked for consistency.
If inconsistent, the initialisation throws an Error and emits a message
about the problem.