eilmer: bug fixes for the solid domain user defined boundary condition.
(1) The temperature at the boundary interface was not being set when
setting the boundary flux directly. We now set the boundary interface
temperature to the value of the neighbouring cell temperature.

(2) There was an off by 1 error in the i,j,k indices (typically used
for accessing cell data) sent to the user's lua script.