grid_utils: update to ugrid_partition to work with gridpro generated su2 files
Previously ugrid_partition assumed that the NELEM and NPOINT sections of the
su2 files were in the order that matched their cell index numbers. For example,
cell 0 was on the first line, cell 1 on the second. Since su2 files have the
cell index listed in each entry this is not strictly neccessary for the file
format, and gridpro su2's apparently do not mind mixing them up. The new
ugrid_parition reading step presizes the cell and node arrays and then slots the
cells and nodes into their indexed position rather than assuming the order
matches.