eilmer: clear imaginary components from complex numbers in flow state and gas state objects.
When performing the complex-step Frechet derivative in the Newton-Krylov accelerator, the flowstate
and gas state values accumulate imaginary components, so we have to start with a clean slate, so to speak.

Originally this was taken care of in an adhoc manner, this commit formalises the procedure with several routines
attached to the relevant objects.