kinetics: change coeff variable from number to double in eval_forward_rate and eval_backward_rate.
The power coefficient for some of the rates can be 0 (or 0+0i for the complex variable code),
our complex math library is returning NaNs when this coefficient is 0+0i.

The quick fix is to change the datatype from number to double.

The coeff variable holds no sensitivity information, and as such is able to be a double type without
effecting the complex-variable differentiation of the code.

TODO: investigate why the result of pow(val, 0+0i) is NaN.