eilmer: refactored residual calculation subroutines to optionally accept a list of cells/faces/vertices.
Forming the numerical Jacobian (for preconditioning in the Newton-Krylov accelerator or for the
adjoint solver) requires the flux to be evaluated on a subset of interfaces (and cells). As a proof
of concept the original numerical Jacobian source code used a local copy of the subroutines used
to evaluate the residual reworked to operate on a given list of cells and faces.

This commit adds this capability to the actual subroutines in the fluidblock object, allowing for
the numerical Jacobian code to reuse these subroutines now. The default behaviour for the
subroutines is retained by a fluidblock operating on all cells and faces if a subset list is not
passed to these routines.