nenzf1d: make use of internal_energy method
so that we include the other modes when asking for the internal energy
of the gas state.

Also, make the comments showing the linear equations consistent with
the formulation for variable area.

And, only compute the Rayleigh-Pitot pressure when the entropy method
is available.  It is not for the 2T air model, at the moment.