eilmer: Loading a FlowSolution object no longer alters the GlobalConfig.
This change fixes a crash when loading a non-shock-fitting solution as the
initial condition for a shock-fitting one. Previously the loaded solution's
grid_motion config variable would get added to the current simulations
GlobalConfig, which the code detected as an error. A similar problem may have
been present with loading another solution with a different grid_type or
different flow_format.