Geometry: compute structured-grid vertex ids with the one function.
Using single_index function only reduces the chance of computing an
inconsistent id value.

Also, now that we have removed the ghost-cell layers from the underlying
storage for structured-grid blocks in the flow solver, we might have
less use for the vtx_id element that is stored with each vertex object.