Eilmer: fix the initialization of Lua interpreters for BCs.
A key change is to initialize interpreters for the BCs after the
gas models for the blocks have been initialized.

That the interpreter for the UDF BC was being initialized before
the block had its gas model initialized was a regression introduced
sometime in the second half of 2020.  Apologies to Marc Ewenz Rocher
for breaking his boundary condition.