eilmer: Added checking on setHistoryPoint function for structured grids.
The setHistoryPoint convenience function available in prep scripts
will now check on the validity of supplied i, j and k indices.
If they are not valid, an error message is printed and the prep stage quits.

This code was prompted by a situation where the user supplied invalid values.
This was not caught until an assert statement in the single_index conversion.
It was hard to discern what caused the issue at that point.