util: getString now throws LuaInputException
In order to report better error messages to users, it's going
to be necessary to throw Exceptions from the Lua helper functions
so that the calling code can provide an error to the user with
more context.

In this commit, I've just changed getString, but I'll need to go
through and do the other service functions as well.