doc/for-developers: Shifted "detailed" examples out of must-have for 4.0 release.
The effort required to build LaTeX documents for each example is going to be large
and push a 4.0 release out to an indeterminate point in the future.
I think we need to approach that documentation goal as a slow burn.

When I chatted to Peter about the time to do one of those examples,
we both estimated they take about a week per example.