Eilmer: start the code for staged preparation of grid and flow files.
The stages of preparation are intended to be:
(1) prepare the geometry and domain definition as a set of one
or more grids.  The grids can be tagged for initial flow condition and
the boundaries will also be tagged with user-defined strings.
(2) using the grid definitions, prepare the flow and configuration files
for the simulation.  Boundary conditions are applied and initial flow
states are written to the flow solution files.

The flag --prep-grid indicates that we want to prepare the grid files
from an input script <job>-grid.lua.

The flag --prep-flow indicates that we want to use the existing grid files
and now write the config and initial flow solution files using the extra
information from the input script <job>-flow.lua.