gasdyn: Complete the state estimates (following thermo updates) in a few places.
Luke and Tamara had fallen over at least one of these bugs once they
used gas models other then the CEA gas model for their shock tunnel
calculations in a Lua script.

The CEA gas model is unusual in that it updates the sound speed whenever
updating the thermo.  The other gas models update the sound-speed with
a separate call in the Lua and Dlang interfaces.