gas: Added binary diffusion calculation for two-temperature gas
I've added a note to check with Nick on this. There's a pressure
term in the binary diffusion calculation. It's not clear to me
if the pressure should be the total gas mixture pressure or
just the pressure of the binary pair involded in the coefficient.
This might require some homework.