examples: Move sphere-nonaka files to a subdirectory marked: deprecated
In preparation for a new version of the sphere-nonaka example,
the present files are moved into a subdirectory called: deprecated-version

Once I'm happy the new veersion is behaving well, these old files will be
removed altogther. The difference is that the new version uses the generalised
CompositeGas and generalised 2-T kinetics. The old version had made use
of the bespoke 2-T air gas model and 2-T air thermochemistry module.