examples: Increase CFL for sphere-nonaka case to 0.75
With the recent improvements to robustness in the code (such
as shock-fitting rework and tight-temperature coupling built-in
for the high-temperature gas work), this case now runs nicely
with a CFL of 0.75, up from 0.25.

Also, I used Nick's advice from the last Eilmer meet up of
running first-order first for shock-fitting simulations.
This simulation runs for a total of 10 body-flow-times (D/u_inf).
There is a delay of 1 bft before shock-fitting starts, and a
delay of 3 bft before switching to the higher order reconstruction.