Eilmer: move the grid-velocities call and the geometry update call into gasdynamic_exlicit_increment_with_moving_grid().
This makes all of the gasdynamic_increment functions look the same and
will allow us to move the logic to selecting the one to use into the
configuration or initialization phase.

The motivation for the code shuffle is to make it easier to introduce
yet another increment scheme to the transient core: an implicit one.