Eilmer: discard MPI_Sync_tasks().
As an experiment to make the code more robust when MPI does not play
well the Dlang exceptions, it was a failure.

For the moment, we need to continue to use e4monitor to catch MPI jobs
that are hung.  This seems to happen most often because an exception
was thrown by one MPI task but not caught and handled well across all
MPI tasks.