Eilmer: allow user-defined source terms to be evaluated at each stage of the explicit update.
Back on the 8th of May (revision 81435aba), I restricted the evaluation
to the start of a time step.

Now, if we set the config option eval_udf_source_terms_at_each_stage
to true,  we can once again have the source terms evaluated
at each stage of the update and retain high-order convergence in time.
The default (false) is to evaluate once only, at the start of the time step.