gasdyn: Bug fix for using finite_wave_dv with a CEAGas object.
In finite_wave_dv, the sound_speed is required early on.
When using a CEAGas, we need to get some CEASavedData in place
BEFORE calling update_sound_speed, so we do that by a dummy call
to update_thermo_from_pT. It's not an entirely dummy call since
we are realying on it to fill in that CEASavedData.

Chris James hit this bug and we worked together to sort it out.