Gasdyn: add dtheta parameter to the conical-shock functions.
For the conical-shock analysis, theta is the angle from the freestream
flow-direction axis to the surface of the cone.
The angular-increment dtheta steps the integrator from the shock-wave
angle to the cone surface, and is a small negative value.

At this point on my workstation, the CEA-backed gas model is not
behaving reliably.  The conical shock demo scripts are running fine,
if we select the LUT gas model.