eilmer: A slightly safer version of the SA turbulence model that should catch both of PJ's issues.
This version of the model reverts the two prior changes and replaces them with
a different expression for "r", the nondimensional argument to the fw function
that supresses the destruction term in the outer region of the boundary layer.
It should correctly handle situations where there are very small strain rates
are similar in magnitude to the complex perturbations, but behave identically
the same in the real number code.