examples: Improved the settings to the Newton-Krylov (fully-coupled implicit) variant of the Mabey flat plate case.
Also set config.diffuse_wall_bcs_on_init to false for both variants, since this provides the fastest convergence for both methods.

On 4 cores of my Intel Xeon E5-2630 CPU, 8ms of simulation time takes 7.5 seconds, for 1212 backward-Euler steps, reaching a residual drop of approximately 1.0e-04.
This compares to the NK's 4.3 seconds for 148 steps, and a residual drop of 1.0e-12.