Eilmer: enable the chemical update integral with the implicit gasdynamic update.
This is for the transient solver.

The default is to do the chemical update in a "split" mode.
To get the chemistry update being integral with the implicit gasdynamic
step, set the configuration option:
config.chemistry_update = "integral"

It is not presently implemented for any of the explicit schemes
for the gasdynamic update.