Eilmer: restore viscous gradient settings for the flow along a cylinder.
We had changed the default setting to calculating gradients at cell
centres because it was generally produced a cleaner viscous simulation,
however, this case seems to show that the old approach of computing
gradients at vertices (using the divergence theorem), does a better job
of resolving the boundary layer.  The profile plot of temperature
through the boundary layer is better matched to the reference (spectral)
result.