ugrid_partition: Added an error message if the user specifies a nDim argument that does not match the dimensionality of the .su2 file.
This change is intended to prevent a bizzare issue (found by a user) where
specifying nDim=1 when partitioning a 2D unstructured grid caused some boundary
faces to be assigned the wrong boundary condition.

For some inexplicable reason, doing this somehow results in a successfully
partitioned grid with no error messages and a runnable simulation (with only a
couple of faces assigned wrong), even though under the hood m2gmetis has
completely screwed up the cell connectivity.

Anyway, it seems reasonable to me that we should check the user's input to make
sure it matches the NDIME line in the su2 file to prevent this.