Eilmer: revisit Hakkinen shock-wave-boundary-layer example.
Make it consistent after the recent change to the default settings
for the spatial derivatives for viscous terms.

Running on helmholtz (Dell Optiplex 990) run time is presently 14152
seconds.  This compares favourably with the October 2019 run on the
same hardware, when the time taken was 15151 seconds.  It is nice that
the code performance has been measurably improved.