Eilmer: catch exceptions that might be thrown by count_invalid_cells.
Not catching the exception when no valid neighbour cells could be found
was allowing the MPI tasks to get out of sync, because the exceptional
flow path did not set a the number of invalid cells for the block.
Now, on catching the exception, we assign the full number of cells
for the block as invalid.  This is seen by the subsequent test across
all MPI tasks.