Eilmer: add a reaction_fraction_schedule to scale the thermochem reactions.
In the decoupled chemistry for the transient code it reduces the time
over which the thermochem is integrated for each gasdynamic step.
For the integral thermochem updates (in the implicit gasdynamic update),
it scales the thermochem source terms that are added to the residual.

This particular commit will almost certainly break Kyle's steady-state
solver but I(PJ) will patch that in the next commit.