eilmer: Added exchange_ghost_cell_turbulent_viscosity, which is called instead of a full exchange when calculating the viscous terms.
Previously, simulations would perform a full exchange of ghost cell data before
computing the viscous terms, to ensure that the mu_t and k_t were up to date on
the boundaries. The new routine exchanges only mu_t and k_t at this point, and
additionally skips the exchange in a laminar flow.