Added the periodic shear layer to the catalog of examples.
Now that the each flow block appears as an individual dataset
in Paraview, it is easier to group them into one multi-block dataset,
merge the blocks, convert cell to point data and then compute the
vorticity without glitches appearing at interblock boundaries.
I find the evolution of the vorticity field interesting to watch.