pitot3: tiny changes to PITOT3 to fix a couple of bugs.
A couple of bugs which I found while trying more examples.

NRST mode was not working as I was checking for the variable
acceleration_tube in locals() not the a variable called 'acceleration_tube'.

when a facility was not specified, it was failing in a couple of places
as it was still looking for a variable which wasn't there.