eilmer: implemented two features to promote robustness for the Newton-Krylov accelerator.
(1) Physicality check: this ensures that, during an update to the conserved variables, the change in
the conserved mass in all cells does not go negative.

(2) Line search: this ensures that, during an update to the conserved variables, the unsteady residual is reduced.

These features are currently experimental and not recommended for wider use.