gas: prep-gas can accept a default collision integral during assembly
When preparing a 2-T gas, the user can supply a selected default
collision integral pair. This might be useful if you have trace
species in which you are NOT concerned about their influence
on the transport properties, and yet you don't have data about their
collision integral behaviour with the main species in the flow.

An example input for prep-gas for a N2-N-He system where we don't have any
information for He interactions with the main species is
shown here. In this case, we select N2:N2 interactions as the default
supplied.

---------
    model = "TwoTemperatureGas"
    species = {'N2', 'N', 'He'}

    options = {
       CI_default = "N2:N2"
    }
---------