nenzf1d: Add a test for Nick's python script.
Oddly, this test fails because the python script does not print
the nozzle exit conditions, but stops just after printing the start
condition for the expansion.  If we run the python script from the
terminal command line, it does the full job.

I don't know why the behaviour changes when it is called from Ruby.