Eilmer: introduced FluidBlock method to set flowstates for the faces.
At yesterday's dev meeting there was some discussion that a new shock
detector needed velocity gradients before the convective-flux calculation.
To enable the gradients to be computed, the flowstates at the faces need
to be in place, however, for the structured-grid blocks,
this was only happening later as part of the convective-flux calculation.
This new method is now called before the shock-detector is called.

Also, clean up a little bit of code around Lachlan's short-cut and
put in a note to ask about the type of S.