eilmer: Further changes to the wall distance method to handle 3D flow better.
Previously, the wall distance algorithm would select the nearest face and then
project onto the face's infinite plane. This caused problems with cells that
are significantly away from a wall but coincidentally are close to the face's
plane, such as at the leading edge of a flat plate, where a few cells of
symmetry are sometimes used. The new algorithm computes the cell's wall
parallel distance in addition to its wall normal distance, and includes the
some amount of parallel distance if it is greater than half the cell length.