eilmer: Only perform diffuse-BC-initialisation at t = 0.0.
Hopefully, this is a fix for Issue #6 in the issue tracker:

Nick Gibbons:
"""
The artificial boundary layer capability seems to diffuse the wall boundaries into the flow, even during a restart, sometimes causing problems.
"""

The new behaviour will only do this when the SimState.time is 0.0.
On restarts, this should not be the case.