Eilmer: apply Nick's fix for MPI_finalize to the transient flow code.
We do want an MPI process that is exiting under normal circumstances
to call MPI_Finalize, however, we do not want that process exiting
under exceptional circumstances to pull the MPI rug from under all of
the other processes.  This should allow the MPI runtime to clean up and
terminate the not-yet-exiting processes.