eilmer: fixed a bug in eq. 71 of the AUSM+up flux calculator...
The M_inf parameter in eq. 71 was not squared as it should be.

This modification has caused the default M_inf value of 0.01 to not
be suitable for the coupled MMS test case - a failure occurs.
To remedy this, M_inf = 0.1 has been set in the Lua input scripts to
allow these test cases to pass.