gas: The vibElecTemperature routine in two_temperature_gas.d once again exits if the starting guess is good.
This commit partially reverts a change made in a545f856, where the routine that
computes Tv from uv was forced to complete a fixed number of iterations in
order to set the complex numbers properly. Unfortunately the check that skipped
iterations was also catching a numerical problem where Tv becomes undefined if
the vibrational energy approaches zero. This commit puts the check back in, but
also manually calculates the imaginary components of T_modes using the analytic
derivatives, so that forced iterations are not required.