Puffin: add a bell-nozzle optimization example.
This example wraps the puffin calculator with an automatic optimization
process and approximates the bell-nozzle example for Eilmer.

Run time for the whole process is about 35 seconds for 73 function
evaluations.  This compares with a little over 6000 seconds for the
optimization using Eilmer.  Individual space-marching calculations
are completed in about 700 milliseconds.