eilmer: enforce positivity of species densities and scale them to sum to the total density in decode_conserved() routine.
The NK-accelerator was observed to exhibit an instability in the boundary layer for simulations that
employed both finite-rate chemistry and mass diffusion. It was discovered that the cause for this
instability was related to the absence of any enforcement of the sum of mass fractions equalling to 1
for the NK-accelerator path of the code.

A fix has been implemented that enforces that the sum of the species densities equals the total density.
This in effect enforces the sum of mass fractions to be 1, with the added benefit of keeping the conserved
quantities vector directly consistent with the primitive variables.

This commit also adds code to prevent negative mass fractions by enforcing postivity of the species densities.