Lua helper function for reading numbers from a table at the top of stack.
This is a variant of the helper function that reads from a table that
is held in the global namespace of the Lua interpreter.
We did not refactor that original function because the D compiler
did not want to pass a shared(double[]) to the new function,
something that is needed in simcore.d.