Eilmer: enhance the handling of user-defined source-terms for Carrie Xie.
Carrie's calculation uses overlaid fluid and solid blocks with the
user-defined source terms used to exchange energy between the gas and
solid on a per-cell interaction.

The code now allows sampling of fluid-cell data when computing source
terms for the solid block and also sampling of the solid-cell data
when computing the fluid-cell source terms.  Also, data sent to the Lua
interpreter when evaluating the solid source terms has been enhanced.
The cell's temperature and its i,j,k indices are now available.

Note that, only structured grids are handled and, in the Lua domain,
the i,j,k indices for the cells fall in range 0 .. ncell, for each
index direction.

The solid blocks still use the old structured-grid indexing internal
to the Dlang code.  Fixing this annoyance is a job for another day.