eilmer: Add u_modes components to total_h in post-processor.
Nils Temme showed me an odd-looking solution in which total
enthalpy was not conserved across the shock. This was a 2-T
simulation and the --add-vars option for total_h did not
include additional internal energy modes. There was a note
in the code to say so, but this is now cleaned up for
multi-temperature simulations.