eilmer: Change to pass-by-value in IO routines for DMD compiler.
When using the DMD compiler, a seg fault would occur in the complex
numbers version of the code in the IO routines. The specific issue
was trying to pass a Vector3 (built with complex numbers) as a
ref const object. The object had values before passing into the
function, but inside the function, any access to the object caused
a seg fault.

The fix was to pass-by-value so that a copy of the struct is made.