eilmer: removed physicality check on total energy and modal energy
This commit removes the failed experiment of calculating the minimum possible internal
energy and modal energy for a gas composition for use in the physicality check.

It is replaced by the more robust method of reducing the relaxation factor until
the decode_conserved() routine returns physically consistent primitive variables.

This commit also defines some magic numbers that existed in the physicality check algorithm.