examples/kinetics: Added air with Helium for 1-T and 2-T
These examples of air with Helium mixture were added to
fulfil a request from Hans Hornung. The examples rely
on access to nearby files in other examples so that
we aren't propagating more copies of reaction schemes
across the code base. A README document explains
how to prepare and use the files.