eilmer: fixed bug regarding the use of the Park limiter with the rhou/rhot thermo interpolators.
The Park limiter requires the spatial gradient of pressure which was not being computed when employing
the rhou/rhot thermo interpolators. This lead to NaNs being thrown when employing the park/hvenkat/hvan_albada
limiter with rhou/rhot interpolators. This commit now ensures that the pressure gradient is calculated
even when employing the rhou/rhot thermo interpolators with the park/hvenkat/hvan_albada limiters.