eilmer: the venkat and van_albada unstructured limiter functions now use nondimensional quantities.
These limiters rely on a smoothing parameter to switch the limiter off in regions where it is not required,
the effect is improved convergence. For this parameter to be effective the limiter functions should be defined
using nondimensional quantities, as noted in Blazek's textbook.

This commit also removes the threshold parameter for the Venkat limiter added in commit 91dba61e, with the
smoothing parameter operating more effectively now, there should be no need for this parameter.