Gas: a bit of mopping-up after the change of GasState to struct.
Since we tightened up a bit of run-time checking when copying the values,
we can no longer start a GasState with arrays suited to a particular
GasModel and then copy data into it from a GasState with arrays that
have differing lengths.

The changes to gas_cwrap.d of the form *gs[i] to *(gs[i]) are to make
my thinking less confused.  There is not change to the operations.