Chicken: added a message to help with debugging the JSON parsing.
Presently, we are having trouble with the JSON parser failing on the
currently-generated config file.  Addind or removing an insignificant
zero digit from either of the floating-point numbers for max_time or
dt_init will allow the parser to complete its work.
The config file is assessed as fine by jsonlint.