Chicken: the gas-dynamic update is done on the GPU.
Presently, the boundary conditions are being applied by the host CPU.
If all of the fluid blocks will fit onto the GPU, we could apply the
boundary conditions on the GPU also and eliminate lots of data copying
at every stage of the update.  The cost would be a fair amount of code.