Chicken: first optimization of the CUDA code: copy less data.
Copy just active cell data or ghost cell data, as needed, at each stage
of the gas-dynamic update.

It might be good compare run-times at this point.
The reference case is the forward-facing-step exercise run with
a CFL value of 0.5 and cell size dx=10mm.

On a Dell Optiplex 990 with a GeForce GTX 1650, the run time is now
55.7s at step 1520. This is down from 77.2s when we were copying all
of the cell data at each stage of the gas-dynamic update.

The OpenMP CPU-only run time is 149.7s on the 4-core Intel i7-2600
in the same Optiplex 990.

From my 2016 workbook notes, page 96 dated 2016-04-01,
and on the same CPU (I think) the Eilmer3 run time was 205s
for a 3-task MPI run at the 1520 step mark.
This old calculation used 2D geometry and predictor-corrector updates.

With Chicken running a full 3D calculation (2 cells in the k-direction)
and using a 3-stage update, we seem to be getting the calculation done
relatively quickly.