lmr5: first commit in public master
Don't panic! v4 of Eilmer is not going away anytime soon.

This commit brings the eilmer 5 work-in-progress code into
the master branch on the public-facing github. This is so
that I can work more closely with the developments in master.

Eilmer v5 is a change at the interface level, but uses almost
all of the core files from Eilmer v4. The migration path
is that new code goes in:

src/lmr

but uses a lot of code from src/eilmer.

There is an example of the new interface in:

examples/lmr/2D/convex-corner