Chicken: have viscous fluxes working
...or, at least, putting the boundary layer in the correct place for
the new flat-plate example.  Note that the grid resolution is not nearly
good enough to properly resolve the boundary-layer profile but the low
resolution simulation does run in about 5 minutes on the Lenovo laptop.

Also, used C++ streams to better read the JSON config file.  The dodgy
C code that I originally used was occasionally putting odd characters
after the closing brace and that would upset the JSON parser.