Chicken: try to activate the per-cell ghost-cell code on the GPU.
As one might guess, this commit makes the aspirational change of
setting the ghost-cell conditions on the GPU but, for the
forward-facing-step example, it is broken.  The CPU-only code works.

Also, despite the previous commit claiming all of the boundary condition
code was moved, we found some more code to move into bcs.cu.