pitot3: added a wrapper to finite_wave_dp to try to make it more reliable
In Pitot3 I used finite_wave_dp a lot for unsteady expansions.

Due to the massive temperature drop it sometimes fails for low pressure conditions.

To deal with this I wrote a function to make no ions versions of gas models
with ions and then I wrote a finite_wave_dp_wrapper function which
breaks the finite_wave_dp up into a set amount of steps
and then if one of the steps fails it can turn ions off for that step
(If we're below a set cutoff temperature) and then it can try again.

It seems to work well! I do use finite_wave_dv in a couple of places,
so I should make a finite_wave_dv wrapper too, but I will get to that,
I will let people mess around with this first.