Chicken: accumulate gradients into local variables.
This follows on from the work that Rob Watt was doing at the recent
hackathon.  It makes quite a difference on the GPU.

On the GTX1650 on helmholtz, the run time for the flat-plate example
at 6880 steps was 96.5 seconds prior to this change.

It is now 65.2 seconds.

The CPU flavour of the code runs in 354.8 seconds on helmholtz.