Chicken: introduce local variables to eval_dUdt
On the GTX 1650, run time for the flat plate example drops to 39.9 sec.
This is a modest gain since merging the kernels but we have yet to try
the same optimization in the viscous terms functions.
For the inviscid simulation of the Kelvin-Helmholtz instability, we see
the run time reduce to 151.9 seconds, down from 227 seconds.