Chicken: add the oblique detonation-wave example from Eilmer3.
There is an Eilmer4 example of the same but the Eilmer3 example
included Rowan's Python code from 2006 for the analytic solution.
A little bit of fiddling and it works with the current gdtk code.
Nice!

At the moment, the example runs with just the ideal gas that is
implemented in chicken.