eilmer: added an option to approximate face derivatives (in solid domains) as a simple average of left/right cell derivatives.
A simple average is typically not advisable since it can lead to odd-even decoupling.

However, I have observed that it can be more robust than our implementation of
the face-tangent augmented face-gradient method for geometrically complex solid domains
where some parts of the mesh are highly warped/skewed.

This is a temporary work-around and should only be used if the face-tangent augmented face-gradient method fails.