Chicken: quick and dirty fix for the isentropic vortex case
Just increase the number of cells in the k-index direction to 3.

When there are only two cells across the k-index direction
and only one interior k-face, the assignment of left_cells[1] and
right_cells[1] to that interior k-face is not correct.
Actually, the problem is in all index directions.

The proper fix (yet to be done) is to be more careful with
the assignment of those cell indices.