PITOT3: added mole fractions to the code output
This wasn't actually too bad in the end, but required a lot of changes
I wrote a code that grabbed the names of every species from the
preset gas models and then got their related molecular weights from
CEA's thermo.inp file and then outputted them to a file.

I think got PITOT3 to import that file and give every facility state
the molecular weight of each species so I could change massf to mole
fractions.

That was pretty easy and then I just had to carry that through
all of the parts of the code...