pitot3: fixed a small bug and added gas states to the output
it turns out that my mole fraction output had a bug when I had to turn
the ions off to get a calculation to solve. The easiest solution
was just to give each FacilityState the full list of species molecular
weights, not just the subset they used to have.

I also added a dictionary of gas states to the output so users
can more easily get access to the gas states, which are otherwise
a bit hidden in other objects. This required changing the scripting
example for that reason.